% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_rhs}
\alias{add_rhs}
\title{Add RHS Parts of Formulas}
\usage{
add_rhs(f1, f2)
}
\arguments{
\item{f1}{formula to extract right side}

\item{f2}{formula to extract right side}
}
\value{
A formula of the form \code{~ rhs(f1) + rhs(f2)}
}
\description{
Add RHS Parts of Formulas
}
\examples{
add_rhs(~ a + b, z ~ c + log(d))
# ~ a + b + c + log(d)
}
