% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{add_splines}
\alias{add_splines}
\title{Add splines to expanded data}
\usage{
add_splines(data_path, period_spline, followup_spline)
}
\arguments{
\item{data_path}{Path to expanded data csv file which has columns \code{for_period} and \code{followup_time}.}

\item{period_spline}{Named list of arguments for \code{ns} for periods spline (column \code{for_period})}

\item{followup_spline}{Named list of arguments for \code{ns} for follow up time (column \code{follow_time})}
}
\value{
Overwrites data_path with new csv with spline basis columns.
}
\description{
Add natural spline basis columns to the expanded data for period and follow up time.
}
