% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{add_splines_df}
\alias{add_splines_df}
\title{Add splines to expanded data in memory}
\usage{
add_splines_df(data, period_spline, followup_spline)
}
\arguments{
\item{data}{A \code{data.table} containing the expanded data.}

\item{period_spline}{Named list of arguments for \code{ns} for periods spline (column \code{for_period})}

\item{followup_spline}{Named list of arguments for \code{ns} for follow up time (column \code{follow_time})}
}
\value{
Returns a data.table with the splines columns added.
}
\description{
Add natural spline basis columns to the expanded data for period and follow up time.
}
