% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{as_formula}
\alias{as_formula}
\title{Coerce to a Formula with RHS only}
\usage{
as_formula(x)
}
\arguments{
\item{x}{A formula or character vector.}
}
\value{
A formula
}
\description{
Coerce to a Formula with RHS only
}
\examples{
as_formula(c("age", "sex"))
}
