% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_monotonic}
\alias{assert_monotonic}
\title{Assert Monotonicity}
\usage{
assert_monotonic(x, increasing = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{increasing}{Test for increasing or decreasing.}
}
\value{
Nothing if check is successful, error otherwise.
}
\description{
Assert Monotonicity
}
\examples{
TrialEmulation:::assert_monotonic(1:3)
TrialEmulation:::assert_monotonic(c(0.02, 0.0187, 0.005), FALSE)
}
\keyword{internal}
