% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{calculate_cum_inc}
\alias{calculate_cum_inc}
\alias{calculate_survival}
\title{Calculate Cumulative Incidence and Survival}
\usage{
calculate_cum_inc(p_mat)

calculate_survival(p_mat)
}
\arguments{
\item{p_mat}{Probability matrix with rows for each subject and follow-up time as the columns.}
}
\value{
A vector containing the cumulative incidence or survival values.
}
\description{
Calculate Cumulative Incidence and Survival
}
\examples{
surv_prob <- matrix(
  c(
    0.1, 0.1, 0.1,
    0.5, 0.2, 0.1
  ),
  nrow = 2,
  byrow = TRUE
)
calculate_cum_inc(surv_prob)
calculate_survival(surv_prob)
}
