% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{calculate_predictions}
\alias{calculate_predictions}
\title{Calculate and transform predictions}
\usage{
calculate_predictions(
  model_matrix,
  treatment_value,
  pred_fun,
  coefs_mat,
  linkinv,
  matrix_n_col
)
}
\arguments{
\item{model_matrix}{Model matrix to multiple with coefficients}

\item{treatment_value}{Value to insert into \code{assigned_treatment} column}

\item{pred_fun}{Function to transform prediction matrix}

\item{coefs_mat}{Matrix of coefficients corresponding to \code{model_matrix}.}

\item{linkinv}{Inverse link function for transforming linear predictor}

\item{matrix_n_col}{Expected number of column after prediction.}
}
\value{
A matrix with transformed predicted values. Number of columns corresponds
to the number of rows of \code{coefs_mat}
}
\description{
Calculate and transform predictions
}
