% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{case_control_func}
\alias{case_control_func}
\title{Case Control Sampling Function}
\usage{
case_control_func(
  period_num,
  data_address,
  n_control = 5,
  data_dir = "~/rds/hpc-work/",
  numCores = NA
)
}
\arguments{
\item{period_num}{Period id}

\item{data_address}{A data.table which is the extended version of input data}

\item{n_control}{Number of controls in the case control sampling Defaults to 5}

\item{data_dir}{Directory to write sampled data csv to}

\item{numCores}{Number of cores for parallel programming}
}
\description{
This function apply case control sampling on the extended data
}
