% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{case_control_sampling}
\alias{case_control_sampling}
\title{Case-control sampling from extended data}
\usage{
case_control_sampling(
  data_dir,
  samples_file = "sample_data.csv",
  min_period,
  max_period,
  n_control,
  numCores = 1
)
}
\arguments{
\item{data_dir}{Directory containing 'sw_data.csv' and "switch_data.csv"}

\item{samples_file}{CSV file name for the  sampled data. Should be the same length as n_control.}

\item{min_period}{First trial period to sample from}

\item{max_period}{Last trial period to sample from}

\item{n_control}{Number of controls for each case to sample}

\item{numCores}{Number of cores used by mclapply for sampling from each trial}
}
\description{
Case-control sampling from extended data
}
