% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{case_util_n}
\alias{case_util_n}
\title{Sample cases utility function}
\usage{
case_util_n(data, n_control = 5, ...)
}
\arguments{
\item{data}{Data to sample from}

\item{n_control}{Number of controls for each case}

\item{...}{ignored}
}
\value{
A data frame with the cases and sampled controls
}
\description{
Sample cases utility function
}
