% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{case_util_p}
\alias{case_util_p}
\title{Sample cases utility function}
\usage{
case_util_p(data, p_control = 0.01, sample_all_periods = FALSE, ...)
}
\arguments{
\item{data}{Data to sample from}

\item{p_control}{Proportion of controls to select}

\item{sample_all_periods}{Sample controls in all periods (TRUE) or only when there is a case in the period (FALSE)?}

\item{...}{ignored}
}
\value{
A data frame with the cases and sampled controls
}
\description{
Sample cases utility function
}
