% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{censor_func}
\alias{censor_func}
\title{Censoring Function in C++}
\arguments{
\item{sw_data}{A dataframe with the columns needed in censoring process}
}
\description{
Artificial censoring C++ function
}
\keyword{internal}
