% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extension.R
\name{check_expand_switch}
\alias{check_expand_switch}
\title{Check Expand Flag After Treatment Switch}
\usage{
check_expand_switch(d)
}
\arguments{
\item{d}{\code{data.frame} containing \code{period_new}}
}
\value{
New \code{expand} values
}
\description{
Check if patients have switched treatment in eligible trials
and set \code{expand = 0}.
}
