% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{check_newdata}
\alias{check_newdata}
\title{Check Data used for Prediction}
\usage{
check_newdata(newdata, model, predict_times)
}
\arguments{
\item{newdata}{new data to predict, or missing.}

\item{model}{glm model object.}

\item{predict_times}{times to predict to add to resulting newdata.}
}
\value{
A \code{newdata} data.frame
}
\description{
Check Data used for Prediction
}
\keyword{internal}
