% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{ci_up_to}
\alias{ci_up_to}
\alias{survival_up_to}
\title{Calculate Cumulative Incidence and Survival}
\usage{
ci_up_to(p_mat)

survival_up_to(p_mat)
}
\arguments{
\item{p_mat}{Probability matrix with rows for each subject and followup time as the columns.}
}
\value{
A vector containing the cumulative incidence or survival values.
}
\description{
Calculate Cumulative Incidence and Survival
}
\examples{
surv_prob <- matrix(
 c(0.1, 0.1, 0.1,
   0.5, 0.2, 0.1),
 nrow = 2,
 byrow = TRUE)
ci_up_to(surv_prob)
}
