% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{cum_inc_up_to}
\alias{cum_inc_up_to}
\alias{survival_up_to}
\title{Calculate Cumulative Incidence and Survival}
\usage{
cum_inc_up_to(p_mat)

survival_up_to(p_mat)
}
\arguments{
\item{p_mat}{Probability matrix with rows for each subject and follow-up time as the columns.}
}
\value{
A vector containing the cumulative incidence or survival values.
}
\description{
Calculate Cumulative Incidence and Survival
}
\examples{
surv_prob <- matrix(
  c(
    0.1, 0.1, 0.1,
    0.5, 0.2, 0.1
  ),
  nrow = 2,
  byrow = TRUE
)
cum_inc_up_to(surv_prob)
}
