% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{data_manipulation}
\alias{data_manipulation}
\title{Data Manipulation Function}
\usage{
data_manipulation(data, use_censor)
}
\arguments{
\item{data}{\code{data.table} to pre-process for weight calculation and extension.}

\item{use_censor}{apply censoring due to treatment switch?}
}
\description{
This function takes the data and all the variables and does the extension preprocessing
}
