% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{data_modelling}
\alias{data_modelling}
\title{Data modelling Function}
\usage{
data_modelling(
  outcomeCov_var = NA,
  outcomeCov = NA,
  outcomeClass = NA,
  model_var = NA,
  first_followup = NA,
  last_followup = NA,
  use_weight = 0,
  run_unweighted_analysis = 0,
  run_weighted_analysis = 1,
  run_p99_analysis = 0,
  run_user_limits_analysis = 0,
  lower_weight = NA,
  upper_weight = NA,
  use_censor = 0,
  check_missing = 0,
  include_followup_time_case = c("linear", "quadratic", "spline"),
  include_expansion_time_case = c("linear", "quadratic", "spline"),
  where_case = NA,
  run_base_model = 1,
  absolutePath,
  numCores = NA,
  glm_function = c("parglm", "glm"),
  use_sample_weights = TRUE
)
}
\arguments{
\item{outcomeCov_var}{List of individual baseline variables used in final model}

\item{outcomeCov}{List of functions of baseline covariates used in final model}

\item{outcomeClass}{Categorical variables used in the final model}

\item{model_var}{List of Variables of interest to be used in final model}

\item{first_followup}{First follow-up period}

\item{last_followup}{Last follow-up period}

\item{use_weight}{Use weights in analysis. If 0 then no weights will be calculated}

\item{run_unweighted_analysis}{Run the final model with no weights when use_weights = 1}

\item{run_weighted_analysis}{Run the final model with original weights}

\item{run_p99_analysis}{Run the final model with truncating the weights at the 1st and 99th percentile}

\item{run_user_limits_analysis}{Run the final model with truncating the weights using user defined limits}

\item{lower_weight}{Use lower weight as minimum possible weight}

\item{upper_weight}{Use upper weight as maximum possible weight}

\item{use_censor}{Use censoring for per-protocol analysis - censor person-times once a person-trial stops taking the initial treatment value}

\item{check_missing}{Check for missing values in final model when use_censor=1 (Not added yet!)}

\item{include_followup_time_case}{The model to include follow up time in outcome model. This has 3 options c("none","linear","quadratic","spline"). The data should contain the corresponding columns: "followup_time","followup_time2" or "followup_base".}

\item{include_expansion_time_case}{The model to include for_period in outcome model. This has 3 options c("none","linear","quadratic","spline").  The data should contain the corresponding columns: "for_period","for_period2" or "period_base".}

\item{where_case}{List of where conditions used in subsetting the data used in final analysis}

\item{run_base_model}{Run the model with no conditions Defaults to 1}

\item{absolutePath}{Direction to where the data for modelling is saved}

\item{numCores}{Number of cores for parallel programming (default value is maximum cores and parallel programming)}

\item{glm_function}{Which glm function to use for the final model from \code{stats} or \code{parglm} packages}

\item{use_sample_weights}{Set model weights to \code{sample_weight * weight}.
data_modelling()}
}
\description{
This function do the modelling.
}
\details{
The class variables paramers (\code{outcomeClass},\code{class_switchn},\code{class_switchd},\code{class_censen},\code{class_censed})
can be given as a character vector which will construct factors using \code{as.factor} or as a named list with the arguments for factor
eg \verb{list(risk_cat=list(levels = c(1,2,3,0), age_cat=list(levels=c(1,2,3),labels=c("50-60","60-70","70+")}
}
