% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expand_func}
\alias{expand_func}
\title{Expand Function}
\arguments{
\item{d}{A dataframe with for_period, period_new and switch_new columns}

\item{first_period}{First period value to start expanding about}
}
\description{
Expand the periods iteratively in C++
}
