% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{expand_helper}
\alias{expand_helper}
\title{Period Expanding Function}
\usage{
expand_helper(y)
}
\arguments{
\item{y}{The data.table with period column}
}
\description{
This function get the data.table with period column and expand it based on it
}
\keyword{internal}
