% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_utils.R
\name{expand_switch}
\alias{expand_switch}
\title{Expand Switch Function}
\usage{
expand_switch(
  id_num,
  data_address,
  outcomeCov_var,
  where_var,
  use_censor,
  followup_spline,
  period_spline,
  maxperiod,
  minperiod,
  lag_p_nosw,
  keeplist,
  data_dir,
  separate_files = FALSE
)
}
\arguments{
\item{id_num}{Vector of Ids}

\item{data_address}{Address for data read with bigmemory}

\item{outcomeCov_var}{A list of individual baseline variables used in final model}

\item{where_var}{Variables used in where conditions used in subsetting the data used in final analysis (where_case), the variables not included in the final model}

\item{use_censor}{Use censoring for per-protocol analysis - censor person-times once a person-trial stops taking the initial treatment value}

\item{followup_spline}{The spline model for followup time when choose "spline" in the include_followup_time_case}

\item{period_spline}{The spline model for for_period when choose "spline" in the include_expansion_time_case}

\item{maxperiod}{Maximum period}

\item{minperiod}{Minimum period}

\item{lag_p_nosw}{when 1 this will set the first weight to be 1 and use p_nosw_d and p_nosw_n at followup-time (t-1) for calculating the weights at followup-time t - can be set to 0 which will increase the maximum and variance of weights (Defaults to 1)}

\item{keeplist}{A list contains names of variables used in final model}

\item{data_dir}{Direction to save data}

\item{separate_files}{Write to one file or one per trial (default FALSE)}
}
\description{
This function performs the data expansion for a chunk of data
}
