% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_sequence.R
\name{expand_trials}
\alias{expand_trials}
\title{Expand trials}
\usage{
expand_trials(object)
}
\arguments{
\item{object}{A \link{trial_sequence} object}
}
\value{
The \link{trial_sequence} \code{object} with a data set containing the full sequence of target trials. The data is
stored according to the options set with \code{\link[=set_expansion_options]{set_expansion_options()}} and especially the \verb{save_to_*} function.
}
\description{
Expand trials
}
