% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extension.R
\name{expand_until_switch}
\alias{expand_until_switch}
\title{Check Expand Flag After Treatment Switch}
\usage{
expand_until_switch(s, n)
}
\arguments{
\item{s}{numeric vector where \code{1} indicates a treatment switch in that period}

\item{n}{length of s}
}
\value{
Vector of indicator values up until first switch.
}
\description{
Check if patients have switched treatment in eligible trials
and set \code{expand = 0}.
}
