% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{f}
\alias{f}
\title{Period Expanding Function}
\usage{
f(y)
}
\arguments{
\item{y}{The data.table with period column}
}
\description{
This function get the data.table with period column and expand it based on it
}
