% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/modelling.R
\name{fit_msm}
\alias{fit_msm}
\alias{fit_msm,trial_sequence-method}
\title{Title}
\usage{
fit_msm(
  object,
  use_sample_weights = TRUE,
  analysis_weights = c("asis", "unweighted", "p99", "weight_limits"),
  weight_limits = c(0, Inf)
)

\S4method{fit_msm}{trial_sequence}(
  object,
  use_sample_weights = TRUE,
  analysis_weights = c("asis", "unweighted", "p99", "weight_limits"),
  weight_limits = c(0, Inf)
)
}
\arguments{
\item{object}{A \code{trial_sequence} object}

\item{use_sample_weights}{logical statement if sample weights should be used, default is \code{TRUE}}

\item{analysis_weights}{a string, one of "asis", "unweighted", "p99", "weight_limits"}

\item{weight_limits}{a numerical vector of length 2}
}
\value{
A modified \code{trial_sequence} object
}
\description{
Title
}
