% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fit_outcome_model}
\alias{fit_outcome_model}
\title{Method for fitting outcome models}
\usage{
fit_outcome_model(object, data, formula)
}
\arguments{
\item{object}{A \code{te_outcome_fitter} object}

\item{data}{\code{data.frame} containing outcomes and covariates as defined in \code{formula}.}

\item{formula}{\code{formula} describing the model.}
}
\value{
An object of class \code{te_outcome_fitted}
}
\description{
Method for fitting outcome models
}
\keyword{internal}
