% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{for_period_func}
\alias{for_period_func}
\title{For_period Feature Function}
\usage{
for_period_func(x)
}
\arguments{
\item{x}{The data.table with id and period columns
for_period_func()}
}
\description{
This function get the data.table with period and id columns and generate the for_period feature
}
\keyword{internal}
