% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{function_that_calculates_bootstrap_CIs}
\alias{function_that_calculates_bootstrap_CIs}
\title{Calculate various types of confidence intervals based on bootstrapping methods}
\usage{
function_that_calculates_bootstrap_CIs(
  object,
  ci_type,
  bootstrap_sample_size = 200,
  predict_times,
  point_estimate,
  pred_fun
)
}
\arguments{
\item{object}{object of clas \code{trial_sequence}}

\item{ci_type}{which CI to generate, options: c('Nonpara. bootstrap', 'LEF outcome', 'LEF both')}

\item{bootstrap_sample_size}{Sample size of bootstrap resampling, default = 200}
}
\value{
2 by x matrix of CI lower and upper bound, x is number of time points
}
\description{
Calculate various types of confidence intervals based on bootstrapping methods
}
