% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{h_extract_baseline}
\alias{h_extract_baseline}
\title{Extract Baseline Observations}
\usage{
h_extract_baseline(trial_file, baseline_file, quiet = TRUE)
}
\arguments{
\item{trial_file}{Path to an expanded trial csv file}

\item{baseline_file}{Path to csv to save baseline observations}

\item{quiet}{Don't print progress messages.}
}
\description{
Extract Baseline Observations
}
\details{
Reads \code{trial_file} and saves the observations with \code{followup_time == 0} to \code{baseline_file} csv.
}
