% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{h_quiet_print}
\alias{h_quiet_print}
\title{Conditional Printing}
\usage{
h_quiet_print(quiet, x, ...)
}
\arguments{
\item{quiet}{(\code{logical}) Messages printed if \code{FALSE}}

\item{x}{Object to print.}

\item{...}{Passed to \code{print} function.}
}
\value{
Value of \code{print(...)} or \code{NULL}
}
\description{
Conditional Printing
}
\examples{
h_quiet_print(quiet = FALSE, "loud hello")
h_quiet_print(quiet = TRUE, "quiet hello")
}
