% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{initiators_ci}
\alias{initiators_ci}
\title{Calculate Cumulative Incidence for Initiators Results}
\usage{
initiators_ci(object, follow_up)
}
\arguments{
\item{object}{result of \code{initiators()}}
}
\value{
Cumulative incidence curves for the two treatment arms.
}
\description{
Calculate Cumulative Incidence for Initiators Results
}
