% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_utils.R
\name{lr}
\alias{lr}
\title{Logistic Regression Function}
\usage{
lr(l)
}
\arguments{
\item{l}{A list contains the data with categorical feature determind if needed and logistic regression formula}
}
\description{
This function get the information needed for performing Logistic Regression for final model
}
