% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/trial_sequence.R
\name{observed_data}
\alias{observed_data}
\alias{observed_data,trial_sequence-method}
\title{Observed Data Accessor and Setter}
\usage{
observed_data(object)

\S4method{observed_data}{trial_sequence}(object)
}
\arguments{
\item{object}{\code{trial_sequence} object}
}
\value{
The observed data from the \verb{@data} slot of \code{object}.
}
\description{
Generic function to observed data
}
\examples{
ts <- trial_sequence("ITT")
ts <- set_data(ts, data_censored)
observed_data(ts)
data.table::set(observed_data(ts), j = "dummy", value = TRUE)
}
