% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/trial_sequence.R
\name{outcome_data}
\alias{outcome_data}
\alias{outcome_data<-}
\alias{outcome_data,trial_sequence-method}
\alias{outcome_data<-,trial_sequence-method}
\title{Outcome Data Accessor and Setter}
\usage{
outcome_data(object)

outcome_data(object) <- value

\S4method{outcome_data}{trial_sequence}(object)

\S4method{outcome_data}{trial_sequence}(object) <- value
}
\arguments{
\item{object}{\code{trial_sequence} object}

\item{value}{\code{data.table} to replace and update in \verb{@outcome_data}}
}
\value{
The \code{object} with updated outcome data
}
\description{
Generic function to outcome data
}
\examples{
ts <- trial_sequence("ITT")
new_data <- data.table::data.table(vignette_switch_data[1:200, ])
new_data$weight <- 1
outcome_data(ts) <- new_data
}
