% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_utils.R
\name{p99_weight}
\alias{p99_weight}
\title{P99 Weights Function}
\usage{
p99_weight(switch_data)
}
\arguments{
\item{switch_data}{The data.table with weight column}
}
\description{
This function truncate the weights of data at the 1st and 99th percentile
}
