% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{predict.RTE_model}
\alias{predict.RTE_model}
\title{Predict Cumulative Incidence with Confidence Intervals}
\usage{
\method{predict}{RTE_model}(
  object,
  newdata,
  predict_times,
  conf_int = TRUE,
  samples = 100,
  type = c("cum_inc", "survival"),
  ...
)
}
\arguments{
\item{object}{Object from \code{\link[=data_modelling]{data_modelling()}} or \code{\link[=initiators]{initiators()}}.}

\item{newdata}{Baseline trial data to predict cumulative incidence or survival for. If \code{newdata} contains
rows with \code{followup_time > 0} these will be removed.}

\item{predict_times}{Follow-up times to predict. Any times given in newdata will be ignored.}

\item{conf_int}{Calculate a confidence interval using coefficient samples from a multivariate normal distribution
based on the robust covariance matrix.}

\item{samples}{The number of samples of the coefficients for prediction models.}

\item{type}{Type of values to calculate. Either cumulative incidence (\code{"cum_inc"}) or survival (\code{"survival"}).}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list of three data frames containing the cumulative incidences for each of the assigned treatment options
and the difference between them.
}
\description{
Predict Cumulative Incidence with Confidence Intervals
}
\examples{
model <- initiators(
  data = trial_example,
  outcome_cov = c("catvarA", "catvarB", "catvarC", "nvarA", "nvarB", "nvarC"),
)

predicted_ci <- predict(model, predict_times = 0:30, samples = 20)

# Plot the cumulative incidence curves for each treatment
plot(predicted_ci[[1]]$followup_time, predicted_ci[[1]]$estimate,
  type = "l",
  xlab = "Follow-up Time", ylab = "Cumulative Incidence"
)
lines(predicted_ci[[1]]$followup_time, predicted_ci[[1]]$`2.5\%`, lty = 2)
lines(predicted_ci[[1]]$followup_time, predicted_ci[[1]]$`97.5\%`, lty = 2)

lines(predicted_ci[[2]]$followup_time, predicted_ci[[2]]$estimate, type = "l", col = 2)
lines(predicted_ci[[2]]$followup_time, predicted_ci[[2]]$`2.5\%`, lty = 2, col = 2)
lines(predicted_ci[[2]]$followup_time, predicted_ci[[2]]$`97.5\%`, lty = 2, col = 2)
legend("topleft", title = "Assigned Treatment", legend = c("0", "1"), col = 1:2, lty = 1)

# Plot the difference in cumulative incidence over follow up
plot(predicted_ci[[3]]$followup_time, predicted_ci[[3]]$estimate,
  type = "l",
  xlab = "Follow-up Time", ylab = "Difference in Cumulative Incidence",
  ylim = c(-0.1, 0.1)
)
lines(predicted_ci[[3]]$followup_time, predicted_ci[[3]]$`2.5\%`, lty = 2)
lines(predicted_ci[[3]]$followup_time, predicted_ci[[3]]$`97.5\%`, lty = 2)

}
