% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data for Trial Emulation}
\usage{
prepare_data(data, setup)
}
\arguments{
\item{data}{A data frame}

\item{setup}{A \code{TE_setup} object from \code{\link[=trial_emulation_setup]{trial_emulation_setup()}}}
}
\value{
A \code{TE_data} object containing a data frame \code{data}
}
\description{
Prepare Data for Trial Emulation
}
