% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.TE_weight_summary}
\alias{print.TE_weight_summary}
\title{Print a Weight Summary Object}
\usage{
\method{print}{TE_weight_summary}(object, full = TRUE, ...)
}
\arguments{
\item{object}{print.TE_weight_summary object}

\item{...}{not used}
}
\description{
Print a Weight Summary Object
}
