% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.TE_weight_summary}
\alias{print.TE_weight_summary}
\title{Print a weight summary object}
\usage{
\method{print}{TE_weight_summary}(x, full = TRUE, ...)
}
\arguments{
\item{x}{print TE_weight_summary object.}

\item{full}{Print full or short summary.}

\item{...}{Arguments passed to \link{print.data.frame}.}
}
\value{
No return value, only for printing.
}
\description{
Print a weight summary object
}
