% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{RandomisedTrialsEmulation-package}
\alias{RandomisedTrialsEmulation}
\alias{RandomisedTrialsEmulation-package}
\title{\code{RandomisedTrialsEmulation} Package}
\description{
\code{RandomisedTrialsEmulation} facilitates preprocessing and analysing observational data
an emulated a randomised trial.
}
\author{
\strong{Maintainer}: Roonak Rezvani \email{roonak.r74@gmail.com} (\href{https://orcid.org/0000-0001-5580-5058}{ORCID})

Authors:
\itemize{
  \item Isaac Gravestock \email{isaac.gravestock@roche.com}
  \item Li Su \email{li.su@mrc-bsu.cam.ac.uk}
}

}
