% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_utils.R
\name{robust_calculation}
\alias{robust_calculation}
\title{Robust Calculation Funstion}
\usage{
robust_calculation(model, data_id)
}
\arguments{
\item{model}{The Logistic Regression model}

\item{data_id}{Values of column id of the data (data\link{, id})}
}
\description{
This function performs the calculation with Robust Standard Errors
}
