% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sampling.R
\name{sample_controls}
\alias{sample_controls}
\alias{sample_controls,trial_sequence-method}
\title{Method to read, subset and sample expanded data}
\usage{
sample_controls(object, period = NULL, subset_condition, p_control = 0.01)

\S4method{sample_controls}{trial_sequence}(object, period = NULL, subset_condition, p_control = 0.01)
}
\arguments{
\item{object}{An object of class \linkS4class{trial_sequence}.}

\item{period}{An integerish vector of non-zero length to select trial period(s) or \code{NULL} (default) to
select all trial periods.}

\item{subset_condition}{A string of length 1 or missing.}

\item{p_control}{Proportion of controls to select.}
}
\value{
A \code{data.frame} of class \code{data.table}.
}
\description{
This method is used on \linkS4class{trial_sequence} objects to read, subset and sample expanded data.
}
