% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_data_prep_sep}
\alias{sample_data_prep_sep}
\title{Sample utility for data_preparation result with separate_files = TRUE}
\usage{
sample_data_prep_sep(
  data_prep,
  p_control,
  sample_all_times,
  use_subset,
  subset_expr
)
}
\description{
Sample utility for data_preparation result with separate_files = TRUE
}
