% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_datastore.R
\name{save_to_datatable}
\alias{save_to_datatable}
\title{Save expanded data as a \code{data.table}}
\usage{
save_to_datatable()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
trial_to_expand <- trial_sequence("ITT") |>
  set_data(data = data_censored) |>
  set_expansion_options(output = save_to_datatable(), chunk_size = 500)
}
\seealso{
Other save_to:
\code{\link[=save_to_csv]{save_to_csv()}},
\code{\link[=save_to_duckdb]{save_to_duckdb()}},
\code{\link[=set_expansion_options]{set_expansion_options()}}
}
\concept{save_to}
