% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_sequence.R
\name{set_censor_weight_model}
\alias{set_censor_weight_model}
\title{Set censoring weight model}
\usage{
set_censor_weight_model(object, ...)
}
\arguments{
\item{object}{trial_sequence.}

\item{censor_event}{string. Name of column containing censoring indicator.}
}
\value{
\code{object} is returned with \verb{@censor_weights} set
}
\description{
Set censoring weight model
}
