% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_sequence.R
\name{set_switch_weight_model,trial_sequence-method}
\alias{set_switch_weight_model,trial_sequence-method}
\title{Set switching weight model}
\usage{
\S4method{set_switch_weight_model}{trial_sequence}(object, numerator, denominator, model_fitter)
}
\arguments{
\item{object}{A \link{trial_sequence} object.}

\item{numerator}{Right hand side formula for the numerator model}

\item{denominator}{Right hand side formula for the denominator model}

\item{model_fitter}{A \link[=te_model_fitter-class]{te_model_fitter} object, such as
\link{stats_glm_logit}}
}
\value{
\code{object} is returned with \verb{@switch_weights} set
}
\description{
Set switching weight model
}
