% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_weights.R
\name{show_weight_models}
\alias{show_weight_models}
\title{Show Weight Model Summaries}
\usage{
show_weight_models(object)
}
\arguments{
\item{object}{A \link{trial_sequence} object after fitting weight models with \code{\link[=calculate_weights]{calculate_weights()}}}
}
\value{
Prints summaries of the censoring models
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
