% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.TE_data_prep}
\alias{summary.TE_data_prep}
\alias{summary.TE_data_prep_sep}
\alias{summary.TE_data_prep_dt}
\alias{summary.TE_msm}
\alias{summary.TE_robust}
\title{Summary methods}
\usage{
\method{summary}{TE_data_prep}(object, ...)

\method{summary}{TE_data_prep_sep}(object, ...)

\method{summary}{TE_data_prep_dt}(object, ...)

\method{summary}{TE_msm}(object, ...)

\method{summary}{TE_robust}(object, ...)
}
\arguments{
\item{object}{Object to print summary}

\item{...}{Additional arguments passed to print methods.}
}
\value{
No value, displays summaries of object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Print summaries of data and model objects produced by \code{TrialEmulation}.
}
