% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{suppressMatchingWarnings}
\alias{suppressMatchingWarnings}
\title{Suppress warnings with matching}
\usage{
suppressMatchingWarnings(
  expr,
  matching = c("probabilities"),
  classes = "warning",
  ...
)
}
\arguments{
\item{expr}{expression}

\item{matching}{a character vector of patterns to pass to \code{\link[=grepl]{grepl()}} for matching the warning messages}

\item{classes}{a character, indicating which classes of warnings should be suppressed}

\item{...}{other arguments to \code{\link[=grepl]{grepl()}}}
}
\description{
Suppress warnings with matching
}
\seealso{
\code{\link[=suppressWarnings]{suppressWarnings()}}
}
