% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{te_censoring}
\alias{te_censoring}
\title{Set up Informative Censoring}
\usage{
te_censoring(
  censored = "censored",
  censor_d_model = ~1,
  censor_n_model = ~1,
  pool_models = FALSE
)
}
\arguments{
\item{censored}{column name with censoring flag}

\item{censor_d_model}{RHS formula denominator model for informative censoring}

\item{censor_n_model}{RHS formula numerator model for informative censoring}

\item{pool_models}{Pool the numerator and denominator models
(\code{FALSE}: split models by previous treatment as in treatment models and
\code{TRUE}: pool all observations together into a single numerator and denominator model)}
}
\description{
Set up Informative Censoring
}
