% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_data.R
\docType{class}
\name{te_data-class}
\alias{te_data-class}
\title{TrialEmulation Data Class}
\description{
TrialEmulation Data Class
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{data.table} object with columns "id", "period",
"treatment", "outcome", "eligible"}
}}

