% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{te_datastore-class}
\alias{te_datastore-class}
\title{Specification for saving expanded data}
\value{
A 'te_datastore' object
}
\description{
This is the parent class for classes which define how the expanded trial data should be stored.
To define a new storage type, a new class should be defined which inherits from \code{te_datastore}. In addition, methods
\link{save_expanded_data} and \code{read_expanded_data} need to be defined for the new class.
}
\section{Slots}{

\describe{
\item{\code{N}}{The number of observations in this data. Initially 0.}
}}

