% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_datastore_csv.R
\docType{class}
\name{te_datastore_csv-class}
\alias{te_datastore_csv-class}
\alias{save_to_csv}
\alias{save_expanded_data,te_datastore_csv-method}
\alias{read_expanded_data,te_datastore_csv-method}
\title{te_datastore_csv, functions and methods}
\usage{
save_to_csv(path)

\S4method{save_expanded_data}{te_datastore_csv}(object, data)

\S4method{read_expanded_data}{te_datastore_csv}(object, period = NULL, subset_condition = NULL)
}
\arguments{
\item{path}{Directory to save CSV files in. Must be empty.}

\item{object}{An object of class \link[=te_datastore-class]{te_datastore} or a child class.}

\item{data}{A data frame containing the expanded trial data. The columns \code{trial_period} and \code{id} are present, which
may be used in methods to save the data in an optimal way, such as with indexes, keys or separate files.}

\item{period}{An integerish vector of non-zero length to select trial period(s) or \code{NULL} (default) to
select all files.}

\item{subset_condition}{A string of length 1 or \code{NULL} (default).}
}
\value{
A 'te_datastore_csv' object.
}
\description{
This method is used internally by \link{expand_trials} to save the data to the "datastore" defined in
\link{set_expansion_options}.
}
\section{Slots}{

\describe{
\item{\code{path}}{path to csv files.}

\item{\code{files}}{names of csv files.}

\item{\code{template}}{data.frame template.}
}}

\examples{
csv_dir <- file.path(tempdir(), "expanded_trials_csv")
dir.create(csv_dir)
csv_datastore <- save_to_csv(path = csv_dir)

trial_to_expand <- trial_sequence("ITT") |>
  set_data(data = data_censored) |>
  set_expansion_options(output = csv_datastore, chunk_size = 500)

# Delete directory after use
unlink(csv_dir)

}
\seealso{
Other save_to: 
\code{\link{save_to_datatable}()},
\code{\link{set_expansion_options}()},
\code{\link{te_datastore_duckdb-class}}
}
\concept{save_to}
