% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{te_model_ex}
\alias{te_model_ex}
\title{Example of a Fitted MSM Model Object

\if{html}{\out{<div class="sourceCode">}}\preformatted{te_model_ex <- trial_msm(
 data = data_subset,
 outcome_cov = c("catvarA", "nvarA"),
 last_followup = 40,
 model_var = "assigned_treatment",
 include_followup_time = ~followup_time,
 include_trial_period = ~trial_period,
 use_sample_weights = FALSE,
 quiet = TRUE,
 glm_function = "glm"
 )
}\if{html}{\out{</div>}}

@seealso \link{te_data_ex}}
\format{
An object of class \code{TE_msm} of length 2.
}
\usage{
te_model_ex
}
\description{
Example of a Fitted MSM Model Object

\if{html}{\out{<div class="sourceCode">}}\preformatted{te_model_ex <- trial_msm(
 data = data_subset,
 outcome_cov = c("catvarA", "nvarA"),
 last_followup = 40,
 model_var = "assigned_treatment",
 include_followup_time = ~followup_time,
 include_trial_period = ~trial_period,
 use_sample_weights = FALSE,
 quiet = TRUE,
 glm_function = "glm"
 )
}\if{html}{\out{</div>}}

@seealso \link{te_data_ex}
}
\keyword{datasets}
