% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_model_fitter.R
\docType{class}
\name{te_model_fitter-class}
\alias{te_model_fitter-class}
\title{Outcome Model Fitter Class}
\description{
This is a virtual class which other outcome model fitter classes should inherit from. Objects of these class exist to
define how the outcome models are fit. They are used for the dispatch of the internal methods \link{fit_outcome_model},
\link{fit_weights_model} and \link{predict}.
}
\seealso{
Other model_fitter:
\code{\link[=parsnip_model]{parsnip_model()}},
\code{\link[=stats_glm_logit]{stats_glm_logit()}}
}
\concept{model_fitter}
