% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_data.R
\docType{class}
\name{te_outcome_data-class}
\alias{te_outcome_data-class}
\title{TrialEmulation Outcome Data Class}
\description{
TrialEmulation Outcome Data Class
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{data.table} object with columns "id", "period",}

\item{\code{n_rows}}{Number of rows}

\item{\code{n_ids}}{Number of IDs}

\item{\code{periods}}{Vector of periods
"treatment", "outcome", "eligible"}
}}

