% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_data.R
\name{te_outcome_data}
\alias{te_outcome_data}
\title{Create te_outcome_data}
\usage{
te_outcome_data(data)
}
\arguments{
\item{data}{A data.table derived from expanded data, containing columns
\code{c("id", "trial_period", "followup_time", "outcome", "weight")}}
}
\value{
A \code{te_outcome_data} object
}
\description{
Create te_outcome_data
}
