% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_outcome_model.R
\docType{class}
\name{te_outcome_fitted-class}
\alias{te_outcome_fitted-class}
\title{Fitted Outcome Model Object}
\description{
Fitted Outcome Model Object
}
\section{Slots}{

\describe{
\item{\code{model}}{list containing fitted model objects.}

\item{\code{summary}}{list of data.frames. Tidy model summaries a la \code{broom()} and \code{glance()}}
}}

