% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_outcome_model.R
\docType{class}
\name{te_outcome_model-class}
\alias{te_outcome_model-class}
\title{Fitted Outcome Model Object}
\description{
Fitted Outcome Model Object
}
\section{Slots}{

\describe{
\item{\code{formula}}{\code{formula} object for the model fitting}

\item{\code{adjustment_vars}}{character. Adjustment variables}

\item{\code{treatment_var}}{Variable used for treatment}

\item{\code{stabilised_weights_terms}}{formula. Adjustment terms from numerator models of stabilised weights. These must be
included in the outcome model.}

\item{\code{adjustment_terms}}{formula. User specified terms to include in the outcome model}

\item{\code{treatment_terms}}{formula. Estimand defined treatment term}

\item{\code{followup_time_terms}}{formula. Terms to model follow up time within an emulated trial}

\item{\code{trial_period_terms}}{formula. Terms to model start time ("trial_period") of an emulated trial}

\item{\code{model_fitter}}{Model fitter object}

\item{\code{fitted}}{list. Saves the model objects}
}}

