% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{te_outcome_model}
\alias{te_outcome_model}
\title{Set up Outcome Model Fitting}
\usage{
te_outcome_model(
  outcome = assigned_treatment ~ .,
  covariate_adjustment = ~1,
  followup_time_adjustment = ~followup_time + I(followup_time^2),
  expansion_time_adjustment = ~for_period + I(for_period^2),
  where_variables,
  use_weight = TRUE,
  glm_function = "glm"
)
}
\arguments{
\item{glm_function}{}
}
\description{
Set up Outcome Model Fitting
}
