% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_parsnip.R
\docType{class}
\name{te_parsnip_model-class}
\alias{te_parsnip_model-class}
\title{Fit Models using parsnip}
\description{
The classes and (internal) methods defined for using \link{parsnip} to fit the weight models.
}
\seealso{
Other model_fitter_classes: 
\code{\link{te_stats_glm_logit-class}}
}
\concept{model_fitter_classes}
\keyword{internal}
