% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{te_sampling}
\alias{te_sampling}
\title{Set up Data Sampling}
\usage{
te_sampling(sample_trials = FALSE, probability = 0.1)
}
\arguments{
\item{sample_trials}{FALSE}

\item{probability}{sampling probability for a control observation}
}
\description{
Set up Data Sampling
}
