% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{te_switching}
\alias{te_switching}
\title{Set Treatment Switching}
\usage{
te_switching(
  switch_n_model = ~1,
  switch_d_model = ~1,
  censor_switchers = TRUE,
  eligible_wts_0 = NA,
  eligible_wts_1 = NA
)
}
\arguments{
\item{eligible_wts_1}{}
}
\description{
Set Treatment Switching
}
