% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{te_variables}
\alias{te_variables}
\title{Set Data Variables}
\usage{
te_variables(
  id = "id",
  period = "period",
  eligible = "eligible",
  treatment = "treatment",
  outcome = "outcome"
)
}
\arguments{
\item{id}{column name of patient IDs}

\item{period}{column name of time period}

\item{eligible}{column name of eligibility flag}

\item{treatment}{column name of treatment received in period}

\item{outcome}{column name of outcome events}
}
\description{
Set Data Variables
}
