% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{te_weight-class}
\alias{te_weight-class}
\title{Weighting Method}
\description{
Weighting Method
}
\section{Slots}{

\describe{
\item{\code{parameters}}{list. Save the parameters that set up the model fitting, eg pooling}

\item{\code{fit}}{list. Saves the model objects or at least summaries if large}
}}

