% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_weights.R
\docType{class}
\name{te_weights_fitted-class}
\alias{te_weights_fitted-class}
\title{Weighting Method}
\description{
Weighting Method
}
\section{Slots}{

\describe{
\item{\code{specification}}{list. The parameters specifying how the model should be fit}

\item{\code{summary}}{list of data.frames. Tidy model summaries a la \code{broom()} and \code{glance()}}

\item{\code{fitted}}{list. Saves the model objects or at least summaries if large}
}}

