% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_weights.R
\docType{class}
\name{te_weights_fitter-class}
\alias{te_weights_fitter-class}
\alias{te_stats_glm_logit-class}
\title{Weight Model Fitter Class}
\description{
This is a virtual class which other weight fitter classes should inherit from. Objects of these
class exist to define how the censoring and switching weight models are fit.
They are used for the dispatch of the method \link{fit_weights_model}
}
\section{Slots}{

\describe{
\item{\code{save_path}}{character. A string specifying a directory to save the weight model objects.}
}}

\seealso{
Other weight_fitter: 
\code{\link{stats_glm_logit}()}
}
\concept{weight_fitter}
