% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{trial_emulation_setup}
\alias{trial_emulation_setup}
\title{Set up Trial Emulation}
\usage{
trial_emulation_setup(
  data,
  variables = te_variables(),
  expansion = te_expansion(),
  treatment_switching = te_switching(),
  censoring = te_censoring(),
  sampling = te_sampling(),
  outcome_model = te_outcome_model(),
  ...
)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{variables}{set the columns in the data frame using \code{\link[=te_variables]{te_variables()}}}

\item{expansion}{set how the data are expanded into trials using \code{\link[=te_expansion]{te_expansion()}}}

\item{treatment_switching}{set how treatment switching is handles}

\item{...}{}
}
\description{
Set up Trial Emulation
}
