% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trial_example}
\alias{trial_example}
\title{Example of longitundinal data for randomised trial emulation}
\format{
A data frame with 48400 rows and 11 variables:
\describe{
\item{id}{patient identifier}
\item{eligible}{eligible for trial start in this period, 1=yes, 0=no}
\item{period}{time period}
\item{outcome}{indicator for outcome in this period, 1=event occurred, 0=no event}
\item{treatment}{indicator for receiving treatment in this period, 1=treatment, 0=no treatment}
\item{catvarA}{A categorical variable relating to treatment and the outcome}
\item{catvarB}{A categorical variable relating to treatment and the outcome}
\item{catvarC}{A categorical variable relating to treatment and the outcome}
\item{nvarA}{A numerical variable relating to treatment and the outcome}
\item{nvarB}{A numerical variable relating to treatment and the outcome}
\item{nvarC}{A numerical variable relating to treatment and the outcome}
}
}
\usage{
trial_example
}
\description{
A dataset containing the treatment, outcomes and other attributes of 503 patients for randomised trial emulation. See \code{vignette("Getting-Started")}
}
\keyword{datasets}
