% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_sequence.R, R/te_stats_glm_logit.R
\docType{class}
\name{trial_sequence-class}
\alias{trial_sequence-class}
\alias{trial_sequence_PP-class}
\alias{trial_sequence_ITT-class}
\alias{trial_sequence_AT-class}
\alias{predict,trial_sequence_PP-method}
\title{Trial Sequence class}
\usage{
\S4method{predict}{trial_sequence_PP}(
  object,
  newdata,
  predict_times,
  conf_int = TRUE,
  samples = 100,
  ci_type = c("sandwich", "Nonpara. bootstrap", "LEF outcome", "LEF both",
    "Jackknife Wald", "Jackknife MVN"),
  type = c("cum_inc", "survival")
)
}
\arguments{
\item{object}{Object to dispatch method on}

\item{newdata}{Baseline trial data that characterise the target trial population that marginal cumulative incidences
or survival probabilities are predicted for.  \code{newdata} must have the same columns and formats of variables as in
the fitted marginal structural model specified in \code{\link[=trial_msm]{trial_msm()}} or \code{\link[=initiators]{initiators()}}. If \code{newdata} contains rows with
\code{followup_time > 0} these will be removed.}

\item{predict_times}{Specify the follow-up visits/times where the marginal cumulative incidences or survival
probabilities are predicted.}

\item{conf_int}{Construct the point-wise 95-percent confidence intervals of cumulative incidences for the target
trial population under treatment and non-treatment and their differences. The default confidence interval
construction methods simulates the parameters in the marginal structural model from a multivariate normal
distribution with the mean equal to the marginal structural model parameter estimates and the variance
equal to the estimated robust covariance matrix.}

\item{samples}{Number of samples used to construct the simulation-based confidence intervals.}

\item{ci_type}{Specify the method used to construct the confidence interval for a per-protocol
estimand (\code{estimand_type = "PP"}) (only available for \link{trial_sequence} object).
\code{'sandwich'}: using the estimated robust sandwich covariance matrix;
\code{'Nonpara. bootstrap'}, \code{'LEF outcome'}, \code{'LEF both'}, \code{'Jackknife Wald'}, \code{'Jackknife MVN'}}

\item{type}{Specify cumulative incidences or survival probabilities to be predicted. Either cumulative incidence
(\code{"cum_inc"}) or survival probability (\code{"survival"}).}
}
\description{
Trial Sequence class
}
\section{Functions}{
\itemize{
\item \code{predict(trial_sequence_PP)}: Predict from the fitted model object via \link{predict} on \code{trial_sequence}

}}
\section{Slots}{

\describe{
\item{\code{data}}{te_data.}

\item{\code{estimand}}{character. Descriptive name of estimand.}

\item{\code{expansion}}{te_expansion}

\item{\code{outcome_model}}{te_outcome_model.}

\item{\code{outcome_data}}{te_outcome_data.}

\item{\code{censor_weight}}{te_weight. Object to define weighting to account for
informative censoring}

\item{\code{censor_weight}}{te_weight. Object to define weighting to account for
informative censoring due to treatment switching}
}}

