% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{TrialEmulation-package}
\alias{TrialEmulation}
\alias{TrialEmulation-package}
\title{\code{TrialEmulation} Package}
\description{
\code{TrialEmulation} facilitates preprocessing and analysing observational data
an emulated a randomised trial.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://causal-lda.github.io/TrialEmulation/}
  \item \url{https://github.com/Causal-LDA/TrialEmulation}
  \item Report bugs at \url{https://github.com/Causal-LDA/TrialEmulation/issues}
}

}
\author{
\strong{Maintainer}: Isaac Gravestock \email{isaac.gravestock@gmail.com} (\href{https://orcid.org/0000-0003-0283-2065}{ORCID})

Authors:
\itemize{
  \item Isaac Gravestock \email{isaac.gravestock@gmail.com} (\href{https://orcid.org/0000-0003-0283-2065}{ORCID})
  \item Li Su \email{li.su@mrc-bsu.cam.ac.uk}
  \item Roonak Rezvani \email{roonak.r74@gmail.com} (\href{https://orcid.org/0000-0001-5580-5058}{ORCID}) (Original package author)
  \item Julia Moesch \email{julia.moesch@roche.com}
  \item Limozin Juliette (\href{https://orcid.org/0000-0002-1140-2350}{ORCID})
}

Other contributors:
\itemize{
  \item Medical Research Council (MRC) [funder]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
