% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_switch_data}
\alias{vignette_switch_data}
\title{Example of expanded longitundinal data for randomised trial emulation}
\format{
A data frame with 1939053 rows and 7 variables:
\describe{
\item{id}{patient identifier}
\item{for_period}{trial start time period}
\item{followup_time}{follow up time within trial}
\item{outcome}{indicator for outcome in this period, 1=event occurred, 0=no event}
\item{treatment}{indicator for receiving treatment in this period, 1=treatment, 0=no treatment}
\item{assigned_treatment}{indicator for assigned treatment at baseline of the trial, 1=treatment, 0=no treatment}
\item{weight}{weights for use with model fitting}
}
}
\usage{
vignette_switch_data
}
\description{
This is the expanded dataset created in the \code{vignette("Getting-Started")} known as \code{switch_data}.
}
\keyword{datasets}
