% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{weight_func_bootstrap}
\alias{weight_func_bootstrap}
\title{Title}
\usage{
weight_func_bootstrap(
  object = trial_pp,
  remodel = TRUE,
  new_coef_sw_d0 = NA,
  new_coef_sw_n0 = NA,
  new_coef_sw_d1 = NA,
  new_coef_sw_n1 = NA,
  new_coef_c_d0 = NA,
  new_coef_c_n0 = NA,
  new_coef_c_d1 = NA,
  new_coef_c_n1 = NA,
  new_coef_c_d = NA,
  new_coef_c_n = NA,
  boot_idx = NA,
  quiet = T,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
