% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_utils.R
\name{weight_lr}
\alias{weight_lr}
\title{Weight Logistic Regression Function}
\usage{
weight_lr(l)
}
\arguments{
\item{l}{A list contains the data, logistic regression formula and name of the categorical variables}
}
\description{
This function get the information needed for performing Logistic Regression in the weight calculation process using parglm
}
