% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{weights.TE_data_prep_dt}
\alias{weights.TE_data_prep_dt}
\alias{weights.TE_data_prep_sep}
\title{Extract Weights}
\usage{
\method{weights}{TE_data_prep_dt}(object, ...)

\method{weights}{TE_data_prep_sep}(object, ...)
}
\arguments{
\item{object}{Object to extract weights from}

\item{...}{Not used.}
}
\value{
Weights extracted from \code{object} as a numeric vector.
}
\description{
Extract Weights
}
