% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{calculate_predictions}
\alias{calculate_predictions}
\title{Calculate and transform predictions}
\usage{
calculate_predictions(
  newdata,
  model,
  treatment_values,
  pred_fun,
  coefs_mat,
  matrix_n_col
)
}
\arguments{
\item{newdata}{New data to predict outcome}

\item{model}{GLM object}

\item{treatment_values}{Named vector of value to insert into \code{assigned_treatment} column}

\item{pred_fun}{Function to transform prediction matrix}

\item{coefs_mat}{Matrix of coefficients corresponding to \code{model_matrix}.}

\item{matrix_n_col}{Expected number of column after prediction.}
}
\value{
A matrix with transformed predicted values. Number of columns corresponds
to the number of rows of \code{coefs_mat}
}
\description{
Calculate and transform predictions
}
\keyword{internal}
