% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_sequence.R
\name{calculate_weights}
\alias{calculate_weights}
\title{Calculate Inverse Probability of Censoring Weights}
\usage{
calculate_weights(object, ...)
}
\arguments{
\item{object}{A \link{trial_sequence} object}

\item{...}{Other arguments used by methods.}

\item{quiet}{Prints model summaries is \code{TRUE}.}
}
\description{
Calculate Inverse Probability of Censoring Weights
}
