% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{case_control_func}
\alias{case_control_func}
\title{Case Control Sampling Function}
\usage{
case_control_func(
  period_num,
  data_address,
  n_control = NULL,
  p_control = NULL,
  sample_all_periods,
  case_util_fun,
  data_dir = "~/rds/hpc-work/",
  name_csv = "sample_data.csv",
  numCores = NA
)
}
\arguments{
\item{period_num}{Period id}

\item{data_address}{A data.table which is the extended version of input data}

\item{n_control}{Number of controls in the case control sampling Defaults to 5}

\item{p_control}{Proportion of controls to select}

\item{sample_all_periods}{Sample controls in all periods (TRUE) or only when there is a case in the period (FALSE)?}

\item{case_util_fun}{The case_util function to use, \code{case_util_p} or \code{case_util_n}.}

\item{data_dir}{Directory to write sampled data csv to}

\item{name_csv}{File name for the sampled data csv file}

\item{numCores}{Number of cores for parallel programming}
}
\description{
This function apply case control sampling on the extended data
}
