% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{case_control_sampling_trials}
\alias{case_control_sampling_trials}
\title{Case-control sampling from extended data in separate trial csvs}
\usage{
case_control_sampling_trials(
  data_dir,
  n_control,
  numCores,
  name_prefix = "cc_sample_",
  infile_pattern = "trial_",
  subset_condition
)
}
\arguments{
\item{data_dir}{Directory containing trial-level extended data}

\item{n_control}{Number of controls to sample per case}

\item{numCores}{Number of cores for parallel processing}

\item{name_prefix}{Name of output csv files for each case-control sample, in the form \code{paste0(name_prefix,seq_along(n_control),"_1x",n_control,".csv")}}

\item{infile_pattern}{Name of trial dataset csv files. This is passed to the \code{pattern} argument of \code{dir()}}

\item{subset_condition}{Expression used to \code{subset} the trial data before sampling}
}
\value{
A vector of file paths of the sampled data
}
\description{
Case-control sampling from extended data in separate trial csvs
}
