% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{case_control_sampling_trials}
\alias{case_control_sampling_trials}
\title{Case-control sampling from extended data in separate trial CSVs}
\usage{
case_control_sampling_trials(
  data_prep,
  n_control = NULL,
  p_control = NULL,
  sample_all_periods,
  subset_condition,
  mc_cores = 1
)
}
\arguments{
\item{data_prep}{Result from \code{\link[=data_preparation]{data_preparation()}}}

\item{n_control}{Number of controls to sample per case}

\item{p_control}{Proportion of controls to select}

\item{sample_all_periods}{Sample controls in all periods (TRUE) or only when there is a case in the period (FALSE)?}

\item{subset_condition}{Expression used to \code{subset} the trial data before sampling}

\item{mc_cores}{Number of cores for parallel processing with mclapply if \code{mc_cores > 1}.}
}
\value{
A \code{data.frame} or a \code{\link[=split]{split()}} \code{data.frame} if  \code{length(n_control) > 1} or \code{length(p_control) > 1}.
}
\description{
Case-control sampling from extended data in separate trial CSVs
}
