% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{CI_up_to}
\alias{CI_up_to}
\title{Calculate Cumulative Incidence}
\usage{
CI_up_to(p_mat, t = ncol(p_mat))
}
\arguments{
\item{p_mat}{Probabilty matrix with rows for each subject and followup time as the columns.}

\item{t}{Number of time periods to use. Must not be more than \code{ncol(p_mat)}}
}
\description{
Calculate Cumulative Incidence
}
