% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extension.R
\name{data_extension}
\alias{data_extension}
\title{Data Extension}
\usage{
data_extension(
  data,
  keeplist,
  outcomeCov_var = NA,
  first_period = NA,
  last_period = NA,
  use_censor = 0,
  where_var = NA,
  data_dir = "~/rds/hpc-work/",
  separate_files = FALSE,
  chunk_size = 200
)
}
\arguments{
\item{data}{A \code{data.table} to be expanded}

\item{keeplist}{A list contains names of variables used in final model}

\item{outcomeCov_var}{A list of individual baseline variables used in final model}

\item{first_period}{First period value to start expanding about}

\item{last_period}{Last period value to expand about}

\item{use_censor}{Use censoring for per-protocol analysis - censor person-times once a person-trial stops taking the
initial treatment value}

\item{where_var}{Variables used in where conditions used in subsetting the data used in final analysis (where_case),
the variables not included in the final model}

\item{data_dir}{Directory to save data}

\item{separate_files}{Save expanded data in separate CSV files for each trial.}

\item{chunk_size}{Number of ids to expand in each chunk}
}
\description{
Expands the longitudinal data into a sequence of trials.
}
