% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extension.R
\name{data_extension}
\alias{data_extension}
\title{Data Extension Function}
\usage{
data_extension(
  data,
  keeplist,
  outcomeCov_var = NA,
  first_period = NA,
  last_period = NA,
  use_censor = 0,
  lag_p_nosw = 1,
  where_var = NA
)
}
\arguments{
\item{data}{A \code{data.frame} or similar}

\item{keeplist}{A list contains names of variables used in final model}

\item{outcomeCov_var}{A list of individual baseline variables used in final model}

\item{first_period}{First period value to start expanding about}

\item{last_period}{Last period value to expand about}

\item{use_censor}{Use censoring for per-protocol analysis - censor person-times once a person-trial stops
taking the initial treatment value}

\item{lag_p_nosw}{when 1 this will set the first weight to be 1 and use \code{p_nosw_d} and \code{p_nosw_n} at
follow-up time (t-1) for calculating the weights at follow-up time t - can be set to 0 which will increase
the maximum and variance of weights (Defaults to 1)}

\item{where_var}{Variables used in where conditions used in subsetting the data used in final
analysis (\code{where_case}), the variables are not included in the final model.}
}
\description{
Expands the longitudinal data into a sequence of trials.
}
