% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{data_extension_parallel}
\alias{data_extension_parallel}
\title{Data Extension in Parallel Function}
\usage{
data_extension_parallel(
  data_address,
  keeplist,
  outcomeCov_var = NA,
  first_period = NA,
  last_period = NA,
  use_censor = 0,
  lag_p_nosw = 1,
  where_var = NA,
  data_dir = "~/rds/hpc-work/",
  numCores = NA,
  chunk_size = 200
)
}
\arguments{
\item{data_address}{Address for data read with bigmemory}

\item{keeplist}{A list contains names of variables used in final model}

\item{outcomeCov_var}{A list of individual baseline variables used in final model}

\item{first_period}{First period value to start expanding about}

\item{last_period}{Last period value to expand about}

\item{use_censor}{Use censoring for per-protocol analysis - censor person-times once a person-trial stops taking the initial treatment value}

\item{lag_p_nosw}{when 1 this will set the first weight to be 1 and use p_nosw_d and p_nosw_n at followup-time (t-1) for calculating the weights at followup-time t - can be set to 0 which will increase the maximum and variance of weights (Defaults to 1)}

\item{where_var}{Variables used in where conditions used in subsetting the data used in final analysis (where_case), the variables not included in the final model}

\item{data_dir}{Direction to save data}

\item{numCores}{Number of cores for parallel programming}

\item{chunk_size}{Number of ids to expand in each chunk
data_extension_parallel()}
}
\description{
This function takes the data and all the variables and expand it using parallel computing
}
