% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{data_modelling}
\alias{data_modelling}
\title{Data modelling Function}
\usage{
data_modelling(
  data,
  outcome_cov = ~1,
  model_var = NA,
  first_followup = NA,
  last_followup = NA,
  use_weight = 0,
  run_unweighted_analysis = 0,
  run_weighted_analysis = 1,
  run_p99_analysis = 0,
  run_user_limits_analysis = 0,
  lower_weight = NA,
  upper_weight = NA,
  use_censor = 0,
  check_missing = 0,
  include_followup_time_case = ~followup_time + I(followup_time^2),
  include_expansion_time_case = ~for_period + I(for_period^2),
  where_case = NA,
  run_base_model = 1,
  numCores = NA,
  glm_function = c("parglm", "glm"),
  use_sample_weights = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} containing all the required columns.}

\item{outcome_cov}{A RHS formula with baseline covariates to adjust in final model}

\item{model_var}{List of Variables of interest to be used in final model.
Derived variables to use in outcome models. Typically \code{assigned_treatment} for ITT and per-protocol,
and \code{dose + dose^2} for as-treated. \code{time_on_regime}? TODO check what else is derived}

\item{first_followup}{First follow-up period}

\item{last_followup}{Last follow-up period}

\item{use_weight}{Use weights in analysis. If 0 then no weights will be calculated}

\item{run_unweighted_analysis}{Run the final model with no weights when use_weights = 1}

\item{run_weighted_analysis}{Run the final model with original weights}

\item{run_p99_analysis}{Run the final model with truncating the weights at the 1st and 99th percentile}

\item{run_user_limits_analysis}{Run the final model with truncating the weights using user defined limits}

\item{lower_weight}{Use lower weight as minimum possible weight}

\item{upper_weight}{Use upper weight as maximum possible weight}

\item{use_censor}{Use censoring for per-protocol analysis - censor person-times once a person-trial stops taking the
initial treatment value}

\item{check_missing}{Check for missing values in final model when use_censor=1 (Not added yet!)}

\item{include_followup_time_case}{The model to include follow up time in outcome model.
This has 3 options c("linear","quadratic","spline")}

\item{include_expansion_time_case}{The model to include for_period in outcome model.
This has 3 options c("linear","quadratic","spline")}

\item{where_case}{List of where conditions used in subsetting the data used in final analysis}

\item{run_base_model}{Run the model with no conditions Defaults to 1}

\item{numCores}{Number of cores for parallel programming (default value is maximum cores and parallel programming)}

\item{glm_function}{Which glm function to use for the final model from \code{stats} or \code{parglm} packages}

\item{use_sample_weights}{Use sample weights in addition to IP weights. \code{data} must contain a column \code{sample_weight}.
The weights used in the model are calculated as \code{weight = weight * sample_weight}.}

\item{quiet}{Don't print progress messages.}
}
\description{
This function do the modelling.
}
\details{
The class variables parameters (\code{outcomeClass},\code{class_switchn},
\code{class_switchd},\code{class_censen},\code{class_censed}) can be given as a character
vector which will construct factors using \code{as.factor} or as a named list
with the arguments for factor e.g.
\verb{list(risk_cat=list(levels = c(1,2,3,0), age_cat=list(levels=c(1,2,3),labels=c("50-60","60-70","70+")}
}
