% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{data_modelling}
\alias{data_modelling}
\title{Fit the Outcome Model}
\usage{
data_modelling(
  data,
  outcome_cov = ~1,
  model_var = NA,
  first_followup = NA,
  last_followup = NA,
  use_weight = 0,
  analysis_weights = c("asis", "unweighted", "p99", "weight_limits"),
  weight_limits = c(0, Inf),
  use_censor = 0,
  include_followup_time = ~followup_time + I(followup_time^2),
  include_expansion_time = ~for_period + I(for_period^2),
  where_case = NA,
  glm_function = c("glm", "parglm"),
  use_sample_weights = TRUE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing all the required columns.}

\item{outcome_cov}{A RHS formula with baseline covariates to adjust in final model}

\item{model_var}{List of Variables of interest to be used in final model.
Derived variables to use in outcome models. Typically \code{assigned_treatment} for ITT and per-protocol,
and \code{dose + dose^2} for as-treated. \code{time_on_regime}? TODO check what else is derived}

\item{first_followup}{First follow-up period}

\item{last_followup}{Last follow-up period}

\item{use_weight}{Use weights in analysis. If 0 then no weights will be calculated.}

\item{analysis_weights}{One of
\itemize{
\item \code{"asis"}: use the weights as calculated
\item \code{"p99"}: truncate weights at the 1st and 99th percentiles
\item \code{"weight_limits"}: truncate weights at the values specified in \code{weight_limits}
\item \code{"unweighted"}: set all analysis weights to 1, even with \code{use_weights = 1}
}}

\item{weight_limits}{Lower and upper limits to truncate weights, given as \code{c(lower, upper)}}

\item{use_censor}{Use censoring for per-protocol analysis - censor person-times once a person-trial stops taking the
initial treatment value}

\item{include_followup_time}{The model to include the follow up time of the trial (\code{followup_time}) in outcome model,
specified as a RHS formula.}

\item{include_expansion_time}{The model to include the trial period (\code{for_period}) in outcome model,
specified as a RHS formula.}

\item{where_case}{List of where conditions used in subsetting the data used in final analysis}

\item{glm_function}{Which glm function to use for the final model from \code{stats} or \code{parglm} packages}

\item{use_sample_weights}{Use sample weights in addition to IP weights. \code{data} must contain a column \code{sample_weight}.
The weights used in the model are calculated as \code{weight = weight * sample_weight}.}

\item{quiet}{Don't print progress messages.}

\item{...}{Additional arguments passed to \code{glm_function}. This may be used to specify initial parameter estimates
or arguments to \code{control}. See \link[stats:glm]{stats::glm}, \link[parglm:parglm]{parglm::parglm} and \code{\link[parglm:parglm.control]{parglm::parglm.control()}} for more information.}
}
\value{
Object of class \code{TE_model} containing
\itemize{
\item \code{model}, a \code{glm} object
\item \code{robust} a list containing a coefficient summary table and the robust covariance \code{matrix}.
}
}
\description{
Sets up the model formulas and data for the pooled logistic regression and
robust variance estimation and fits the models.
}
\details{
The class variables parameters (\code{outcomeClass},\code{class_switchn},
\code{class_switchd},\code{class_censen},\code{class_censed}) can be given as a character
vector which will construct factors using \code{as.factor} or as a named list
with the arguments for factor e.g.
\verb{list(risk_cat=list(levels = c(1,2,3,0), age_cat=list(levels=c(1,2,3),labels=c("50-60","60-70","70+")}
}
