% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{data_preparation}
\alias{data_preparation}
\title{Data Preparation Function}
\usage{
data_preparation(
  data,
  id = "id",
  period = "period",
  treatment = "treatment",
  outcome = "outcome",
  eligible = "eligible",
  outcome_cov = ~1,
  model_var = NULL,
  switch_n_cov = ~1,
  switch_d_cov = ~1,
  first_period = NA,
  last_period = NA,
  use_weight = 0,
  use_censor = 0,
  check_missing = 0,
  cense = NA,
  pool_cense = 0,
  cense_d_cov = ~1,
  cense_n_cov = ~1,
  include_followup_time_case = ~followup_time + I(followup_time^2),
  include_expansion_time_case = ~for_period + I(for_period^2),
  include_regime_length = FALSE,
  eligible_wts_0 = NA,
  eligible_wts_1 = NA,
  lag_p_nosw = 1,
  where_var = NULL,
  data_dir,
  numCores = 1,
  chunk_size = 500,
  separate_files = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} containing all the required columns.}

\item{id}{Name of the data column for id feature Defaults to id}

\item{period}{Name of the data column for period feature Defaults to period}

\item{treatment}{Name of the data column for treatment feature Defaults to treatment}

\item{outcome}{Name of the data column for outcome feature Defaults to outcome}

\item{eligible}{Indicator of whether or not an observation is eligible to be expanded about Defaults to eligible}

\item{outcome_cov}{A RHS formula with baseline covariates to adjust in final model}

\item{model_var}{List of Variables of interest to be used in final model.
Derived variables to use in outcome models. Typically \code{assigned_treatment} for ITT and per-protocol,
and \code{dose + dose^2} for as-treated. \code{time_on_regime}? TODO check what else is derived}

\item{switch_n_cov}{A RHS formula for modelling probability of switching treatment. Used in the numerator of weight
calculation.}

\item{switch_d_cov}{A RHS formula for modelling probability of switching treatment. Used in the denominator of weight
calculation.}

\item{first_period}{First period value to start expanding about}

\item{last_period}{Last period value to expand about}

\item{use_weight}{Use weights in analysis. If 0 then no weights will be calculated}

\item{use_censor}{Use censoring for per-protocol analysis - censor person-times once a person-trial stops taking the
initial treatment value}

\item{check_missing}{Check for missing values in final model when use_censor=1 (Not added yet!)}

\item{cense}{Censoring variable}

\item{pool_cense}{Pool the numerator and denominator models (0: split models by previous treatment Am1 = 0 and
Am1 = 1 as in treatment models and 1: pool all observations together into a single numerator and denominator model)
Defaults to 0}

\item{cense_d_cov}{A RHS formula for modelling probability of being censored. Used in the numerator of weight
calculation.}

\item{cense_n_cov}{A RHS formula for modelling probability of being censored. Used in the denominator of weight
calculation.}

\item{include_followup_time_case}{The model to include follow up time in outcome model.
This has 3 options c("linear","quadratic","spline")}

\item{include_expansion_time_case}{The model to include for_period in outcome model.
This has 3 options c("linear","quadratic","spline")}

\item{include_regime_length}{If defined as 1 a new variable (time_on_regime) is added to dataset.
This variable stores the duration of time that the patient has been on the current treatment value}

\item{eligible_wts_0}{Eligibility criteria used in weights for model condition Am1 = 0}

\item{eligible_wts_1}{Eligibility criteria used in weights for model condition Am1 = 1}

\item{lag_p_nosw}{When 1 this will set the first weight to be 1 and use p_nosw_d and p_nosw_n at followup-time (t-1)
for calculating the weights at followup-time t - can be set to 0 which will increase the maximum and variance of
weights (Defaults to 1)}

\item{where_var}{List of variables used in where conditions used in subsetting the data used in final analysis
(where_case), the variables not included in the final model}

\item{data_dir}{Direction to save data}

\item{numCores}{Number of cores for parallel programming (default value is maximum cores and parallel programming)}

\item{chunk_size}{Number of ids to process at once for the chunk expansion (default 500). Larger chunk_sizes
may be faster but require more memory.}

\item{separate_files}{Write to one file or one per trial (default FALSE)}

\item{quiet}{Don't print progress messages.}
}
\description{
This function prepare the data for modelling.
}
\details{
The class variables parameters (\code{outcomeClass},\code{class_switchn},\code{class_switchd},\code{class_censen},
\code{class_censed}) can be given as a character vector which will construct factors using \code{as.factor} or as a named list
with the arguments for factor e.g.
\verb{list(risk_cat=list(levels = c(1,2,3,0), age_cat=list(levels=c(1,2,3),labels=c("50-60","60-70","70+")}
}
