% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_utils.R
\name{expand}
\alias{expand}
\title{Expand Function}
\usage{
expand(
  sw_data,
  outcomeCov_var,
  where_var,
  use_censor,
  maxperiod,
  minperiod,
  lag_p_nosw,
  keeplist,
  data_dir,
  separate_files
)
}
\arguments{
\item{sw_data}{datatable to expand}

\item{outcomeCov_var}{A list of individual baseline variables used in final model}

\item{where_var}{Variables used in where conditions used in subsetting the data used in final analysis (where_case), the variables not included in the final model}

\item{use_censor}{Use censoring for per-protocol analysis - censor person-times once a person-trial stops taking the initial treatment value}

\item{maxperiod}{Maximum period}

\item{minperiod}{Minimum period}

\item{lag_p_nosw}{when 1 this will set the first weight to be 1 and use p_nosw_d and p_nosw_n at followup-time (t-1) for calculating the weights at followup-time t - can be set to 0 which will increase the maximum and variance of weights (Defaults to 1)}

\item{keeplist}{A list contains names of variables used in final model}

\item{data_dir}{Direction to save data}

\item{separate_files}{Write to one file or one per trial}
}
\description{
This function performs the data expansion for a given dataset
}
