% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/modelling.R
\name{fit_msm}
\alias{fit_msm}
\alias{fit_msm,trial_sequence-method}
\title{Fit the marginal structural model for the sequence of emulated trials}
\usage{
fit_msm(
  object,
  use_sample_weights = TRUE,
  analysis_weights = c("asis", "unweighted", "p99", "weight_limits"),
  weight_limits = c(0, Inf)
)

\S4method{fit_msm}{trial_sequence}(
  object,
  use_sample_weights = TRUE,
  analysis_weights = c("asis", "unweighted", "p99", "weight_limits"),
  weight_limits = c(0, Inf)
)
}
\arguments{
\item{object}{A \code{trial_sequence} object}

\item{use_sample_weights}{logical statement if sample weights should be used, default is \code{TRUE}}

\item{analysis_weights}{a string, one of \verb{"asis"``, }"unweighted"\verb{, `"p99"}, \code{"weight_limits"}.}

\item{weight_limits}{a numeric vector of length 2 . If \verb{analysis_weights = "weight_limits} this specifies the upper
and lower limits for truncating the weights.

Before the outcome marginal structural model can be fit, the outcome model must be specified with
\code{\link[=set_outcome_model]{set_outcome_model()}} and the data must be expanded into the trial sequence with \code{\link[=expand_trials]{expand_trials()}}.

The model is fit based on the \code{model_fitter} specified in \link{set_outcome_model} using the internal \code{fit_outcome_model}
method.}
}
\value{
A modified \code{trial_sequence} object with updated \code{outcome_model} slot.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
trial_seq_object <- trial_sequence("ITT") |>
  set_data(data_censored) |>
  set_outcome_model(
    adjustment_terms = ~age_s,
    followup_time_terms = ~ stats::poly(followup_time, degree = 2)
  ) |>
  set_expansion_options(output = save_to_datatable(), chunk_size = 500) |>
  expand_trials() |>
  load_expanded_data()

fit_msm(trial_seq_object)
}
