% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fit_outcome_model}
\alias{fit_outcome_model}
\title{Method for fitting outcome models}
\usage{
fit_outcome_model(object, data, formula, weights = NULL)
}
\arguments{
\item{object}{A \code{te_outcome_fitter} object}

\item{data}{\code{data.frame} containing outcomes and covariates as defined in \code{formula}.}

\item{formula}{\code{formula} describing the model.}

\item{weights}{\code{numeric} vector of weights.}
}
\value{
An object of class \code{te_outcome_fitted}
}
\description{
Method for fitting outcome models
}
\examples{
fit_outcome_model
}
\keyword{internal}
