% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiators.R
\name{initiators}
\alias{initiators}
\title{Initiators Analysis}
\usage{
initiators(
  data,
  id = "id",
  period = "period",
  treatment = "treatment",
  outcome = "outcome",
  eligible = "eligible",
  outcome_cov = ~1,
  model_var = NULL,
  switch_n_cov = ~1,
  switch_d_cov = ~1,
  first_period = NA,
  last_period = NA,
  first_followup = NA,
  last_followup = NA,
  use_weight = 0,
  save_weight_models = c("tidy", "object", "no"),
  run_unweighted_analysis = 0,
  run_weighted_analysis = 1,
  run_p99_analysis = 0,
  run_user_limits_analysis = 0,
  lower_weight = NA,
  upper_weight = NA,
  use_censor = 0,
  check_missing = 0,
  cense = NA,
  pool_cense = 0,
  cense_d_cov = ~1,
  cense_n_cov = ~1,
  include_followup_time_case = ~followup_time + I(followup_time^2),
  include_expansion_time_case = ~for_period + I(for_period^2),
  include_regime_length = 0,
  eligible_wts_0 = NA,
  eligible_wts_1 = NA,
  lag_p_nosw = 1,
  where_var = NULL,
  where_case = NA,
  run_base_model = 1,
  data_dir,
  numCores = NA,
  glm_function = "glm",
  quiet = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} containing all the required columns.}

\item{id}{Name of the data column for id feature Defaults to id}

\item{period}{Name of the data column for period feature Defaults to period}

\item{treatment}{Name of the data column for treatment feature Defaults to treatment}

\item{outcome}{Name of the data column for outcome feature Defaults to outcome}

\item{eligible}{Indicator of whether or not an observation is eligible to be expanded about Defaults to eligible}

\item{outcome_cov}{A RHS formula with baseline covariates to adjust in final model}

\item{model_var}{List of Variables of interest to be used in final model.
Derived variables to use in outcome models. Typically \code{assigned_treatment} for ITT and per-protocol,
and \code{dose + dose^2} for as-treated. \code{time_on_regime}? TODO check what else is derived}

\item{switch_n_cov}{A RHS formula for modelling probability of switching treatment. Used in the numerator of weight
calculation.}

\item{switch_d_cov}{A RHS formula for modelling probability of switching treatment. Used in the denominator of weight
calculation.}

\item{first_period}{First period value to start expanding about}

\item{last_period}{Last period value to expand about}

\item{first_followup}{First follow-up period}

\item{last_followup}{Last follow-up period}

\item{use_weight}{Use weights in analysis. If 0 then no weights will be calculated}

\item{save_weight_models}{Save weight models. Specify \code{"tidy"} to save summaries produced by \code{broom} package,
\code{"object"} to save \code{glm} model objects, or \code{"no"} to save nothing.}

\item{run_unweighted_analysis}{Run the final model with no weights when use_weights = 1}

\item{run_weighted_analysis}{Run the final model with original weights}

\item{run_p99_analysis}{Run the final model with truncating the weights at the 1st and 99th percentile}

\item{run_user_limits_analysis}{Run the final model with truncating the weights using user defined limits}

\item{lower_weight}{Use lower weight as minimum possible weight}

\item{upper_weight}{Use upper weight as maximum possible weight}

\item{use_censor}{Use censoring for per-protocol analysis - censor person-times once a person-trial stops taking the
initial treatment value}

\item{check_missing}{Check for missing values in final model when use_censor=1 (Not added yet!)}

\item{cense}{Censoring variable}

\item{pool_cense}{Pool the numerator and denominator models (0: split models by previous treatment Am1 = 0 and
Am1 = 1 as in treatment models and 1: pool all observations together into a single numerator and denominator model)
Defaults to 0}

\item{cense_d_cov}{A RHS formula for modelling probability of being censored. Used in the numerator of weight
calculation.}

\item{cense_n_cov}{A RHS formula for modelling probability of being censored. Used in the denominator of weight
calculation.}

\item{include_followup_time_case}{The model to include \code{followup_time} in outcome model, specified as a RHS formula.}

\item{include_expansion_time_case}{The model to include \code{for_period} in outcome model, specified as a RHS formula.}

\item{include_regime_length}{If defined as 1 a new variable (time_on_regime) is added to dataset.
This variable stores the duration of time that the patient has been on the current treatment value}

\item{eligible_wts_0}{Eligibility criteria used in weights for model condition Am1 = 0}

\item{eligible_wts_1}{Eligibility criteria used in weights for model condition Am1 = 1}

\item{lag_p_nosw}{When 1 this will set the first weight to be 1 and use p_nosw_d and p_nosw_n at followup-time (t-1)
for calculating the weights at followup-time t - can be set to 0 which will increase the maximum and variance of
weights (Defaults to 1)}

\item{where_var}{List of variables used in where conditions used in subsetting the data used in final analysis
(where_case), the variables not included in the final model}

\item{where_case}{List of where conditions used in subsetting the data used in final analysis}

\item{run_base_model}{Run the model with no conditions Defaults to 1}

\item{data_dir}{Direction to save data}

\item{numCores}{Number of cores for parallel programming (default value is maximum cores and parallel programming)}

\item{glm_function}{Which glm function to use for the final model from \code{stats} or \code{parglm} packages}

\item{quiet}{Don't print progress messages.}
}
\description{
An all-in-one analysis using a sequence of target trials. This provides a simplified
interface to the main working functions \code{\link[=data_preparation]{data_preparation()}} and \code{\link[=data_modelling]{data_modelling()}}.
}
