% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiators.R
\name{initiators}
\alias{initiators}
\title{Initiators Analysis}
\usage{
initiators(
  data,
  id = "id",
  period = "period",
  treatment = "treatment",
  outcome = "outcome",
  eligible = "eligible",
  outcome_cov = ~1,
  model_var = NULL,
  switch_n_cov = ~1,
  switch_d_cov = ~1,
  first_period = NA,
  last_period = NA,
  first_followup = NA,
  last_followup = NA,
  use_weight = 0,
  save_weight_models = FALSE,
  analysis_weights = c("asis", "unweighted", "p99", "weight_limits"),
  weight_limits = c(0, Inf),
  use_censor = 0,
  check_missing = 0,
  cense = NA,
  pool_cense = 0,
  cense_d_cov = ~1,
  cense_n_cov = ~1,
  include_followup_time_case = ~followup_time + I(followup_time^2),
  include_expansion_time_case = ~for_period + I(for_period^2),
  include_regime_length = 0,
  eligible_wts_0 = NA,
  eligible_wts_1 = NA,
  lag_p_nosw = 1,
  where_var = NULL,
  where_case = NA,
  data_dir,
  glm_function = "glm",
  quiet = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} containing all the required columns.}

\item{id}{Name of the data column for id feature Defaults to id}

\item{period}{Name of the data column for period feature Defaults to period}

\item{treatment}{Name of the data column for treatment feature Defaults to treatment}

\item{outcome}{Name of the data column for outcome feature Defaults to outcome}

\item{eligible}{Indicator of whether or not an observation is eligible to be expanded about Defaults to eligible}

\item{outcome_cov}{A RHS formula with baseline covariates to adjust in final model}

\item{model_var}{List of Variables of interest to be used in final model.
Derived variables to use in outcome models. Typically \code{assigned_treatment} for ITT and per-protocol,
and \code{dose + dose^2} for as-treated. \code{time_on_regime}? TODO check what else is derived}

\item{switch_n_cov}{A RHS formula for modelling probability of switching treatment. Used in the numerator of weight
calculation.}

\item{switch_d_cov}{A RHS formula for modelling probability of switching treatment. Used in the denominator of weight
calculation.}

\item{first_period}{First period value to start expanding about}

\item{last_period}{Last period value to expand about}

\item{first_followup}{First follow-up period}

\item{last_followup}{Last follow-up period}

\item{use_weight}{Use weights in analysis. If 0 then no weights will be calculated.}

\item{save_weight_models}{Save weight models objects in \code{data_dir}.}

\item{analysis_weights}{One of
\itemize{
\item \code{"asis"}: use the weights as calculated
\item \code{"p99"}: truncate weights at the 1st and 99th percentiles
\item \code{"weight_limits"}: truncate weights at the values specified in \code{weight_limits}
\item \code{"unweighted"}: set all analysis weights to 1, even with \code{use_weights = 1}
}}

\item{weight_limits}{Lower and upper limits to truncate weights, given as \code{c(lower, upper)}}

\item{use_censor}{Use censoring for per-protocol analysis - censor person-times once a person-trial stops taking the
initial treatment value}

\item{check_missing}{Check for missing values in final model when use_censor=1 (Not added yet!)}

\item{cense}{Censoring variable}

\item{pool_cense}{Pool the numerator and denominator models (0: split models by previous treatment Am1 = 0 and
Am1 = 1 as in treatment models and 1: pool all observations together into a single numerator and denominator model)
Defaults to 0}

\item{cense_d_cov}{A RHS formula for modelling probability of being censored. Used in the numerator of weight
calculation.}

\item{cense_n_cov}{A RHS formula for modelling probability of being censored. Used in the denominator of weight
calculation.}

\item{include_followup_time_case}{The model to include \code{followup_time} in outcome model, specified as a RHS formula.}

\item{include_expansion_time_case}{The model to include \code{for_period} in outcome model, specified as a RHS formula.}

\item{include_regime_length}{If defined as 1 a new variable (time_on_regime) is added to dataset.
This variable stores the duration of time that the patient has been on the current treatment value}

\item{eligible_wts_0}{Eligibility criteria used in weights for model condition Am1 = 0}

\item{eligible_wts_1}{Eligibility criteria used in weights for model condition Am1 = 1}

\item{lag_p_nosw}{When 1 this will set the first weight to be 1 and use p_nosw_d and p_nosw_n at followup-time (t-1)
for calculating the weights at followup-time t - can be set to 0 which will increase the maximum and variance of
weights (Defaults to 1)}

\item{where_var}{List of variables used in where conditions used in subsetting the data used in final analysis
(where_case), the variables not included in the final model}

\item{where_case}{List of where conditions used in subsetting the data used in final analysis}

\item{data_dir}{Direction to save data}

\item{glm_function}{Which glm function to use for the final model from \code{stats} or \code{parglm} packages}

\item{quiet}{Don't print progress messages.}
}
\description{
An all-in-one analysis using a sequence of target trials. This provides a simplified
interface to the main working functions \code{\link[=data_preparation]{data_preparation()}} and \code{\link[=data_modelling]{data_modelling()}}.
}
\details{
If \code{model_var = NULL} the package will add some terms to the outcome model:
\itemize{
\item if \code{use_censor = 0} and \code{use_weight = 0}, an as-treated analysis will be done the outcome model will have
\code{~ dose + I(dose^2)} terms added
\item if \code{use_censor = 1}, a per-protocol analysis will be done with an \code{~assigned_treatment} term added
\item if \code{use_censor = 0} and \code{use_weight = 1}, an intention to treat analysis will be done with an
\code{~assigned_treatment} term added
}
}
