% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/trial_sequence.R
\name{internal-methods}
\alias{internal-methods}
\alias{set_data,trial_sequence,data.frame-method}
\alias{set_expansion_options,trial_sequence-method}
\alias{expand_trials,trial_sequence-method}
\title{Internal Methods}
\usage{
\S4method{set_data}{trial_sequence,data.frame}(
  object,
  data,
  censor_at_switch,
  id = "id",
  period = "period",
  treatment = "treatment",
  outcome = "outcome",
  eligible = "eligible"
)

\S4method{set_expansion_options}{trial_sequence}(
  object,
  output,
  chunk_size = 0,
  first_period = 0,
  last_period = Inf,
  censor_at_switch
)

\S4method{expand_trials}{trial_sequence}(object)
}
\arguments{
\item{object}{A \linkS4class{trial_sequence} object}

\item{data}{A \code{data.frame} containing all the required variables in the person-time format, i.e., the
<U+2018>long<U+2019> format.}

\item{id}{Name of the variable for identifiers of the individuals. Default is <U+2018>id<U+2019>.}

\item{period}{Name of the variable for the visit/period. Default is <U+2018>period<U+2019>.}

\item{treatment}{Name of the variable for the treatment indicator at that visit/period. Default is
<U+2018>treatment<U+2019>.}

\item{outcome}{Name of the variable for the indicator of the outcome event at that visit/period. Default is
<U+2018>outcome<U+2019>.}

\item{eligible}{Name of the variable for the indicator of eligibility for the target trial at that visit/period.
Default is <U+2018>eligible<U+2019>.}
}
\description{
Various S4 methods which are not directly for use by users.
}
\keyword{internal}
