% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{predict.RTE_model}
\alias{predict.RTE_model}
\title{Predict Cumulative Incidence or Survival with Confidence Intervals}
\usage{
\method{predict}{RTE_model}(
  object,
  newdata,
  followup_time,
  type = c("cum_inc", "surv"),
  conf_int = TRUE,
  samples = 100
)
}
\arguments{
\item{object}{Object from \code{\link[=data_modelling]{data_modelling()}}}

\item{newdata}{Baseline trial data to predict cumulative incidence or survival for.}

\item{followup_time}{Follow-up times to predict. Any times given in newdata will be ignored.}

\item{type}{Predict cumulative incidence ("cum_inc") or survival ("surv").}

\item{conf_int}{Calculate a confidence interval using coefficient samples from a multivariate normal distribution
based on the robust covariance matrix.}

\item{samples}{The number of samples of the coefficients for prediction models.}
}
\value{
A list of predictions
}
\description{
Predict Cumulative Incidence or Survival with Confidence Intervals
}
