% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{predict_survival}
\alias{predict_survival}
\title{Predict Cumulative Incidence}
\usage{
predict_survival(model, predict_followup)
}
\arguments{
\item{predict_followup}{a vector of follow up times to predict values for.}

\item{object}{the result from \code{initiators} or a \code{glm} object.}
}
\value{
a list with estimated cumulative incidence curves for \code{assigned_treatment}
equal to 0 and 1.
}
\description{
Predict Cumulative Incidence
}
