% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{predict_survival}
\alias{predict_survival}
\title{Predict Cumulative Incidence}
\usage{
predict_survival(object, model, predict_followup, newdata)
}
\arguments{
\item{object}{The result from \code{initiators}. Either object or model must be specified.}

\item{model}{A \code{glm} object.}

\item{predict_followup}{a vector of follow up times to predict values for.}

\item{newdata}{\code{data.frame} to predict survival for. Extracted from \code{model}/\code{object} if not provided.}
}
\value{
a list with estimated cumulative incidence curves for \code{assigned_treatment}
equal to 0 and 1.
}
\description{
Predict Cumulative Incidence
}
