% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.TE_weight_summary}
\alias{print.TE_weight_summary}
\title{Print a Weight Summary Object}
\usage{
\method{print}{TE_weight_summary}(x, full = TRUE, digits = 4, ...)
}
\arguments{
\item{x}{print.TE_weight_summary object.}

\item{full}{Print full or short summary.}

\item{...}{Arguments from other methods. Not used.}
}
\description{
Print a Weight Summary Object
}
