% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/te_datastore.R
\name{read_expanded_data}
\alias{read_expanded_data}
\alias{read_expanded_data,te_datastore_csv-method}
\title{Method to read expanded data}
\usage{
read_expanded_data(object, period = NULL)

\S4method{read_expanded_data}{te_datastore_csv}(object, period = NULL)
}
\arguments{
\item{object}{An object of class \link{te_datastore_csv}.}

\item{period}{An integerish vector of non-zero length to select trial period(s) or \code{NULL} (default) to select all files.}
}
\value{
A \code{data.frame} of class \code{data.table}.
}
\description{
This method is used on \link{te_datastore_csv} objects to read selected csv-files and combine them into one \code{data.table}.
}
