% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{robust_calculation}
\alias{robust_calculation}
\title{Robust Calculation Function}
\usage{
robust_calculation(model, data_id)
}
\arguments{
\item{model}{The logistic regression model object.}

\item{data_id}{Values of id column of the data (ie \code{data[, id]}) to identify clusters.}
}
\value{
A list with elements \code{summary}, a table with the model summary using the
robust variance estimates, and \code{matrix}, the \code{sandwich} covariance matrix.
}
\description{
This function performs the calculation of robust standard errors based on
variances estimated using \code{\link[sandwich:vcovCL]{sandwich::vcovCL}}.
}
