% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sampling.R
\name{sample_expanded_data}
\alias{sample_expanded_data}
\alias{sample_expanded_data,data.frame-method}
\title{Internal method to sample expanded data}
\usage{
sample_expanded_data(object, p_control)

\S4method{sample_expanded_data}{data.frame}(object, p_control)
}
\arguments{
\item{p_control}{Proportion of controls to select}
}
\value{
A \code{data.frame} of class \code{data.table}.
}
\description{
Internal method to sample expanded data
}
