% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_datastore.R
\name{save_to_datatable}
\alias{save_to_datatable}
\title{Save expanded data as a \code{data.table}}
\usage{
save_to_datatable()
}
\description{
Save expanded data as a \code{data.table}
}
\examples{
trial_to_expand <- trial_sequence("ITT") |>
  set_data(data = data_censored) |>
  set_expansion_options(output = save_to_datatable(), chunk_size = 500)
}
\seealso{
Other save_to: 
\code{\link{set_expansion_options}()},
\code{\link{te_datastore_csv-class}},
\code{\link{te_datastore_duckdb-class}}
}
\concept{save_to}
