% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{select_data_cols}
\alias{select_data_cols}
\title{Select Data Columns}
\usage{
select_data_cols(
  data,
  id = "id",
  period = "period",
  treatment = "treatment",
  outcome = "outcome",
  eligible = "eligible",
  eligible_wts_0 = NA,
  eligible_wts_1 = NA,
  outcomeCov_var = NA,
  cov_switchn = NA,
  cov_switchd = NA,
  cov_censed = NA,
  cov_censen = NA,
  cense = NA,
  where_var = NA
)
}
\arguments{
\item{data}{A \code{data.frame} containing all the required columns}

\item{id}{Name of the \code{data} column for id feature Defaults to "id"}

\item{period}{Name of the \code{data} column for period feature Defaults to "period"}

\item{treatment}{Name of the \code{data} column for treatment feature Defaults to "treatment"}

\item{outcome}{Name of the \code{data} column for outcome feature Defaults to "outcome"}

\item{eligible}{Name of \code{data} column for indicator of whether or not an observation is eligible for a trial.
Defaults to "eligible".}

\item{eligible_wts_0}{Eligibility criteria used in weights for model condition Am1 = 0}

\item{eligible_wts_1}{Eligibility criteria used in weights for model condition Am1 = 1}

\item{outcomeCov_var}{List of individual baseline variables used in final model}

\item{cov_switchn}{Covariates to be used in logistic model for switching probabilities for numerator model}

\item{cov_switchd}{Covariates to be used in logistic model for switching probabilities for denominator model}

\item{cov_censed}{Covariates to be used in logistic model for censoring weights for denominator model}

\item{cov_censen}{Covariates to be used in logistic model for censoring weights for nominator model}

\item{cense}{Censoring variable}

\item{where_var}{Variables used in where conditions used in subsetting the data used in final analysis (where_case),
the variables not included in the final model.}
}
\description{
Select the required columns from the data and rename
}
