% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{select_data_cols}
\alias{select_data_cols}
\title{Select Data Columns}
\usage{
select_data_cols(data, args)
}
\arguments{
\item{data}{A \code{data.frame} containing all the required columns}

\item{args}{List of arguments from \code{data_preparation}}
}
\value{
A data.table with the required columns
}
\description{
Select the required columns from the data and rename
}
\keyword{internal}
