% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_sequence.R
\name{set_data}
\alias{set_data}
\title{Set the trial data}
\usage{
set_data(object, data, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing all the required variables in the
person-time format, i.e., the <U+2018>long<U+2019> format.}

\item{trial_sequence}{A \linkS4class{trial_sequence} object}

\item{id}{Name of the variable for identifiers of the individuals. Default is
<U+2018>id<U+2019>.}

\item{period}{Name of the variable for the visit/period. Default is <U+2018>period<U+2019>.}

\item{treatment}{Name of the variable for the treatment indicator at that
visit/period. Default is <U+2018>treatment<U+2019>.}

\item{outcome}{Name of the variable for the indicator of the outcome event at
that visit/period. Default is <U+2018>outcome<U+2019>.}

\item{eligible}{Name of the variable for the indicator of eligibility for the
target trial at that visit/period. Default is <U+2018>eligible<U+2019>.}
}
\value{
A \code{trial_sequence} object with data
}
\description{
Set the trial data
}
\examples{
data(trial_example)
trial_sequence("ITT") |>
  set_data(
    data = trial_example,
    id = "id",
    period = "period",
    eligible = "eligible",
    treatment = "treatment"
  )

}
