% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_sequence.R
\name{set_switch_weight_model,trial_sequence-method}
\alias{set_switch_weight_model,trial_sequence-method}
\title{Set switching weight model}
\usage{
\S4method{set_switch_weight_model}{trial_sequence}(object, numerator, denominator, model_fitter)
}
\arguments{
\item{object}{trial_sequence.}

\item{model_fitter}{}
}
\value{
\code{object} is returned with \verb{@switch_weights} set
}
\description{
Set switching weight model
}
