% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_weights.R
\name{stats_glm_logit}
\alias{stats_glm_logit}
\title{Fit weight models using \code{stats::glm}}
\usage{
stats_glm_logit(save_path)
}
\arguments{
\item{save_path}{A string specifying a directory to save the weight model objects.}
}
\value{
An object of class \code{te_stats_glm_logit} inheriting from \linkS4class{te_weights_fitter} which is used for
dispatching methods for the fitting models.
}
\description{
Specify that the pool logistic regression models should be fit using \link[stats:glm]{stats::glm} with \code{family = binomial(link = "logit")}.
}
\examples{
stats_glm_logit(tempdir())
}
\seealso{
Other weight_fitter: 
\code{\link{te_weights_fitter-class}}
}
\concept{weight_fitter}
