% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{summary.TE_data_prep}
\alias{summary.TE_data_prep}
\alias{summary.TE_data_prep_sep}
\alias{summary.TE_data_prep_dt}
\title{Summary methods for data_preparation objects}
\usage{
\method{summary}{TE_data_prep}(object, ...)

\method{summary}{TE_data_prep_sep}(object, ...)

\method{summary}{TE_data_prep_dt}(object, ...)
}
\arguments{
\item{object}{Result from data_preparation()}

\item{...}{}
}
\description{
Summary methods for data_preparation objects
}
