% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.TE_data_prep}
\alias{summary.TE_data_prep}
\alias{summary.TE_data_prep_sep}
\alias{summary.TE_data_prep_dt}
\alias{summary.TE_model}
\alias{summary.TE_robust}
\title{Summary methods}
\usage{
\method{summary}{TE_data_prep}(object, ...)

\method{summary}{TE_data_prep_sep}(object, ...)

\method{summary}{TE_data_prep_dt}(object, ...)

\method{summary}{TE_model}(object, ...)

\method{summary}{TE_robust}(object, ...)
}
\arguments{
\item{object}{Object to print summary}

\item{...}{Additional arguments passed to print methods.}
}
\description{
Print summaries of data and model objects produced by \code{TrialEmulation}.
}
