% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_datastore_csv.R
\docType{class}
\name{te_datastore_csv-class}
\alias{te_datastore_csv-class}
\alias{save_expanded_data,te_datastore_csv-method}
\alias{read_expanded_data,te_datastore_csv-method}
\title{te_datastore_csv, functions and methods}
\usage{
\S4method{save_expanded_data}{te_datastore_csv}(object, data)

\S4method{read_expanded_data}{te_datastore_csv}(object, period = NULL, subset_condition = NULL)
}
\arguments{
\item{object}{An object of class \link[=te_datastore-class]{te_datastore} or a child class.}

\item{data}{A data frame containing the expanded trial data. The columns \code{trial_period} and \code{id} are present, which
may be used in methods to save the data in an optimal way, such as with indexes, keys or separate files.}

\item{period}{An integerish vector of non-zero length to select trial period(s) or \code{NULL} (default) to
select all files.}

\item{subset_condition}{A string of length 1 or \code{NULL} (default).}
}
\value{
A 'te_datastore_csv' object.
}
\description{
This method is used internally by \link{expand_trials} to save the data to the "datastore" defined in
\link{set_expansion_options}.
}
\section{Slots}{

\describe{
\item{\code{path}}{path to csv files.}

\item{\code{files}}{names of csv files.}

\item{\code{template}}{data.frame template.}
}}

\examples{
temp_dir <- tempfile("csv_dir_")
dir.create(temp_dir)
datastore <- save_to_csv(temp_dir)
data(vignette_switch_data)
save_expanded_data(datastore, vignette_switch_data[1:200, ])

# delete after use
unlink(temp_dir, recursive = TRUE)
}
\keyword{internal}
