% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_datastore_duckdb.R
\docType{class}
\name{te_datastore_duckdb-class}
\alias{te_datastore_duckdb-class}
\title{te_datastore_duckdb, functions and methods}
\value{
A 'te_datastore_duckdb' object.
}
\description{
te_datastore_duckdb, functions and methods
}
\section{Slots}{

\describe{
\item{\code{path}}{Path to the duckdb file containing the data.}

\item{\code{table}}{.}

\item{\code{con}}{S4 object of class duckdb_connection.}
}}

