% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_outcome_model.R
\docType{class}
\name{te_outcome_model-class}
\alias{te_outcome_model-class}
\title{Fitted Outcome Model Object}
\description{
Fitted Outcome Model Object
}
\section{Slots}{

\describe{
\item{\code{formula}}{\code{formula} object for the model fitting}

\item{\code{treatment_var}}{character. The treatment variable}

\item{\code{adjustment_vars}}{character. Adjustment variables}

\item{\code{model_fitter}}{Model fitter object}

\item{\code{fitted}}{list. Saves the model objects}
}}

