% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_parsnip.R
\docType{class}
\name{te_parsnip_model-class}
\alias{te_parsnip_model-class}
\alias{fit_weights_model,te_parsnip_model-method}
\title{Fit Models using parsnip}
\usage{
\S4method{fit_weights_model}{te_parsnip_model}(object, data, formula, label)
}
\arguments{
\item{object}{The object determining which method should be used, containing any slots containing user defined
parameters.}

\item{data}{\code{data.frame} containing outcomes and covariates as defined in \code{formula}.}

\item{formula}{\code{formula} describing the model.}

\item{label}{A short string describing the model.}
}
\description{
The classes and (internal) methods defined for using parsnip to fit the weight models.
}
\section{Functions}{
\itemize{
\item \code{fit_weights_model(te_parsnip_model)}: Fit the weight models object via \link{calculate_weights} on \code{trial_sequence}

}}
\section{Slots}{

\describe{
\item{\code{model_spec}}{A model specification defined with the \code{parsnip} package.}
}}

\seealso{
Other model_fitter_classes: 
\code{\link{te_stats_glm_logit-class}}
}
\concept{model_fitter_classes}
\keyword{internal}
