% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{trial_emulation_setup}
\alias{trial_emulation_setup}
\title{Set up Trial Emulation}
\usage{
trial_emulation_setup(
  data,
  variables = te_variables(),
  expansion = te_expansion(),
  treatment_switching = te_switching(),
  censoring = te_censoring(),
  sampling = te_sampling(),
  outcome_model = te_outcome_model(),
  ...
)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{variables}{set the columns in the data frame using \code{\link[=te_variables]{te_variables()}}}

\item{expansion}{set how the data are expanded into trials using \code{\link[=te_expansion]{te_expansion()}}}

\item{treatment_switching}{set how treatment switching weighted and censored using \code{\link[=te_switching]{te_switching()}}}

\item{censoring}{set how informative censoring is weighted using \code{\link[=te_censoring]{te_censoring()}}}

\item{sampling}{set how case-control sampling is done for large datasets \code{\link[=te_sampling]{te_sampling()}}}

\item{outcome_model}{specify the outcome model using \code{\link[=te_outcome_model]{te_outcome_model()}}}

\item{...}{}
}
\description{
Set up Trial Emulation
}
