% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_sequence.R
\docType{class}
\name{trial_sequence-class}
\alias{trial_sequence-class}
\alias{trial_sequence_PP-class}
\alias{trial_sequence_ITT-class}
\alias{trial_sequence_AT-class}
\title{Trial Sequence class}
\description{
Trial Sequence class
}
\section{Slots}{

\describe{
\item{\code{data}}{te_data.}

\item{\code{estimand}}{character. Descriptive name of estimand.}

\item{\code{expansion}}{te_expansion}

\item{\code{outcome_model}}{te_outcome_model.}

\item{\code{censor_weight}}{te_weight. Object to define weighting to account for
informative censoring}

\item{\code{censor_weight}}{te_weight. Object to define weighting to account for
informative censoring due to treatment switching}
}}

