% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_sequence.R
\name{trial_sequence}
\alias{trial_sequence}
\title{Create a sequence of emulated target trials object}
\usage{
trial_sequence(estimand, ...)
}
\arguments{
\item{estimand}{The name of the estimand for this analysis, either one of \code{"ITT"}, \code{"PP"}, \code{"AT"} for
intention-to-treat, per-protocol, as-treated estimands respectively, or the name of a class extending
\linkS4class{trial_sequence}}

\item{...}{Other parameters used when creating object}
}
\value{
An estimand specific trial sequence object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
trial_sequence("ITT")

}
