% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{TrialEmulation-package}
\alias{TrialEmulation}
\alias{TrialEmulation-package}
\title{\code{TrialEmulation} Package}
\description{
\code{TrialEmulation} facilitates preprocessing and analysing observational data
an emulated a randomised trial.
}
\author{
\strong{Maintainer}: Isaac Gravestock \email{isaac.gravestock@roche.com}

Authors:
\itemize{
  \item Roonak Rezvani \email{roonak.r74@gmail.com} (\href{https://orcid.org/0000-0001-5580-5058}{ORCID}) (Original package author)
  \item Li Su \email{li.su@mrc-bsu.cam.ac.uk}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item MRC Biostatistics Unit, University of Cambridge [copyright holder, funder]
}

}
