% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{weight_func}
\alias{weight_func}
\title{Weight Calculation Function}
\usage{
weight_func(
  sw_data,
  cov_switchn = NA,
  model_switchn = NA,
  class_switchn = NA,
  cov_switchd = NA,
  model_switchd = NA,
  class_switchd = NA,
  eligible_wts_0 = NA,
  eligible_wts_1 = NA,
  cense = NA,
  pool_cense = 0,
  cov_censed = NA,
  model_censed = NA,
  class_censed = NA,
  cov_censen = NA,
  model_censen = NA,
  class_censen = NA,
  include_regime_length = 0,
  numCores = NA,
  save_dir,
  quiet = FALSE
)
}
\arguments{
\item{sw_data}{A data.table}

\item{cov_switchn}{List of covariates to be used in logistic model for switching probabilities for numerator model}

\item{model_switchn}{List of models (functions) to use the covariates from cov_switchn}

\item{class_switchn}{Class variables used in logistic model for nominator model}

\item{cov_switchd}{List of covariates to be used in logistic model for switching probabilities for denominator model}

\item{model_switchd}{List of models (functions) to use the covariates from cov_switchd}

\item{class_switchd}{Class variables used in logistic model for denominator model}

\item{eligible_wts_0}{Eligibility criteria used in weights for model condition Am1 = 0}

\item{eligible_wts_1}{Eligibility criteria used in weights for model condition Am1 = 1}

\item{cense}{Censoring variable}

\item{pool_cense}{Pool the numerator and denominator models (0: split models by previous treatment Am1 = 0 and Am1 = 1 as in treatment models and 1: pool all observations together into a single numerator and denominator model) Defaults to 0}

\item{cov_censed}{List of covariates to be used in logistic model for censoring weights in denominator model}

\item{model_censed}{List of models (functions) to use the covariates from cov_censed}

\item{class_censed}{Class variables used in censoring logistic regression in denominator model}

\item{cov_censen}{List of covariates to be used in logistic model for censoring weights in numerator model}

\item{model_censen}{List of models (functions) to use the covariates from cov_censen}

\item{class_censen}{Class variables used in censoring logistic regression in numerator model}

\item{include_regime_length}{If defined as 1 a new variable (time_on_regime) is added to dataset - This variable stores the duration of time that the patient has been on the current treatment value}

\item{numCores}{Number of cores for parallel programming}

\item{save_dir}{Directory to save tidy weight model summaries in as 'weight_models.rda'}

\item{quiet}{Don't print progress messages.}
}
\description{
This function performs the calculation for weight of the data
}
