% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{weight_func_bootstrap}
\alias{weight_func_bootstrap}
\title{Weighting for bootstrap samples}
\usage{
weight_func_bootstrap(
  object,
  remodel = TRUE,
  new_coef_sw_d0 = NA,
  new_coef_sw_n0 = NA,
  new_coef_sw_d1 = NA,
  new_coef_sw_n1 = NA,
  new_coef_c_d0 = NA,
  new_coef_c_n0 = NA,
  new_coef_c_d1 = NA,
  new_coef_c_n1 = NA,
  new_coef_c_d = NA,
  new_coef_c_n = NA,
  boot_idx,
  quiet = TRUE,
  glm_function = "glm",
  ...
)
}
\arguments{
\item{object}{trial_sequence class object with weight and outcome models fitted}

\item{remodel}{indicates whether user requires refitting the weight models, default = TRUE}

\item{new_coef_sw_d0}{new model coefficients for switching weight denominator model in previous treatment = 0}

\item{new_coef_sw_n0}{new model coefficients for switching weight denominator model in previous treatment = 0}

\item{new_coef_sw_d1}{new model coefficients for switching weight denominator model in previous treatment = 0}

\item{new_coef_sw_n1}{new model coefficients for switching weight denominator model in previous treatment = 0}

\item{new_coef_c_d0}{new model coefficients for switching weight denominator model in previous treatment = 0}

\item{new_coef_c_n0}{new model coefficients for switching weight denominator model in previous treatment = 0}

\item{new_coef_c_d1}{new model coefficients for switching weight denominator model in previous treatment = 0}

\item{new_coef_c_n1}{new model coefficients for switching weight denominator model in previous treatment = 0}

\item{new_coef_c_d}{new model coefficients for switching weight denominator model in previous treatment = 0}

\item{new_coef_c_n}{new model coefficients for switching weight denominator model in previous treatment = 0}

\item{boot_idx}{vector containing bootstrap sample of patient ids}

\item{quiet}{indicates whether function messages are printed or not, default = TRUE}

\item{glm_function}{glm function to be used for weight model fitting, default is "glm"}

\item{...}{}
}
\value{
List containing
\describe{
\item{model}{Outcome data with new refitted/recalculated weights from given bootstrap sample}
\item{switch_models}{a list containing a summary table of regression models for switching weights}
\item{cense_models}{a list containing a summary table of regression models for censoring weights}
}
}
\description{
Weighting for bootstrap samples
}
