% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_utils.R
\name{weight_lr}
\alias{weight_lr}
\title{Weight Logistic Regression Function}
\usage{
weight_lr(data, formula, class_var)
}
\arguments{
\item{data}{data}

\item{formula}{model formula for \code{parglm}}

\item{class_var}{categorical variables to be converted to factors}
}
\description{
This function get the information needed for performing Logistic Regression in the weight calculation process using parglm
}
